% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.spectre.R
\name{add.new.data}
\alias{add.new.data}
\title{Add new data to Spectre object}
\usage{
add.new.data(spectre_obj, dat, dat_name, metadata = NULL)
}
\arguments{
\item{spectre_obj}{A Spectre object.}

\item{dat}{A data.table to add to the object.}

\item{dat_name}{The name of the data.}

\item{metadata}{A named List. DEFAULT = NULL.
Additional information (metadata) about the data, including its description.
Each element in the list must be associated with a name.}
}
\value{
An updated Spectre object.
}
\description{
Add a new data to the list in Spectre object.
}
\examples{
dat <- create.spectre.object("cell_id")
x <- Spectre::demo.asinh
x$cell_id <- paste0("cell_", seq(1, nrow(x)))
dat <- add.new.data(dat, x, "cyto_asinh", metadata = list("description"= "test"))
dat
}
