% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_to_data_table}
\alias{add_to_data_table}
\title{Add columns in one data table (dat_to_add) to another (dat).
Replace columns with the same name}
\usage{
add_to_data_table(dat, dat_to_add)
}
\description{
Add columns in one data table (dat_to_add) to another (dat).
Replace columns with the same name
}
