% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.files.R
\name{adjust.files}
\alias{adjust.files}
\title{adjust.files}
\usage{
adjust.files(dat, colnames.table)
}
\arguments{
\item{dat}{List of data.tables, typically one per 'sample'}

\item{colnames.table}{A data.table of column names. First column contains the current names, second column contains the new names to update to.}
}
\description{
adjust.files
}
