% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alpha_list}
\alias{alpha_list}
\title{Matrix of predicted alpha diversity in each cell.}
\format{vector.}
\usage{
alpha_list
}
\description{
Matrix of predicted alpha diversity in each cell.
}
\keyword{datasets}
