% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.files.R
\name{annotate.files}
\alias{annotate.files}
\title{annotate.files - Add column of new values or keywords to samples within a list of sample dataframes}
\usage{
See \url{https://sydneycytometry.org.au/spectre} for usage instructions and vignettes.
}
\arguments{
\item{x}{The list of sample dataframes. No default.}

\item{col.name}{Name of the new column. No Default.}

\item{d}{List of values or keywords to add to each. No default.}

\item{assignments}{A list designating which new values or keywords are added to each sample. No default.}

\item{add.num}{Logical, add a 'numbered' version of the new values or keywords. Defaults to FALSE.}
}
\value{
a list (called data.list) of dataframes -- one file per dataframe -- with the a column of the new values or keywords added.
}
\description{
This function allows you to add a column of new values or keywords to the rows (cells) of each sample dataframe within a list.
}
\examples{
annotate.files()

}
\references{
\url{https://sydneycytometry.org.au/spectre}
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
