% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_commonness_error.R
\name{calc_commonness_error}
\alias{calc_commonness_error}
\title{calc_commonness_error}
\usage{
calc_commonness_error(x, objective_matrix)
}
\arguments{
\item{x}{Results object from run_optimization_min_conf.}

\item{objective_matrix}{Matrix from (modeled) alpha-diversity and Bray-Curtis dissimilarity}
}
\value{
vector
}
\description{
Calculate commonness error
}
\details{
Calculate mean absolute commonness error (MAE_c) and relative commonness error in percentage (RCE).
}
