% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_commonness.R
\name{calculate_commonness}
\alias{calculate_commonness}
\title{Function calculate the common species pair-wise matrix (i.e. the target) 
I have just made this very simply and it is likely this is an excellent place to try and improve efficiency. 
FEELING LIKE I AM MISSING OBVIOUS IMPROVEMENTS HERE}
\usage{
calculate_commonness(alpha, beta)
}
\arguments{
\item{alpha}{predicted alpha diverisity.}

\item{beta}{pairwise matrix of predicted dissimilarity.}
}
\value{
matrix of common species between each site
}
\description{
Function calculate the common species pair-wise matrix (i.e. the target) 
I have just made this very simply and it is likely this is an excellent place to try and improve efficiency. 
FEELING LIKE I AM MISSING OBVIOUS IMPROVEMENTS HERE
}
