% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_solution_commonness.R
\name{calculate_solution_commonness}
\alias{calculate_solution_commonness}
\title{Function calculates the commonness matrix for site-species solutions
Uses equation: Cij = sum(total_species(i) * total_species(j))
AGAIN I THINK THIS IS A PLACE WHERE WE CAN IMPROVE EFFICIENCY}
\usage{
calculate_solution_commonness(solution_matrix)
}
\arguments{
\item{solution_matrix}{predicted species-site matrix.}
}
\value{
matrix of common species between each site
}
\description{
Function calculates the commonness matrix for site-species solutions
Uses equation: Cij = sum(total_species(i) * total_species(j))
AGAIN I THINK THIS IS A PLACE WHERE WE CAN IMPROVE EFFICIENCY
}
