% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_solution_commonness.R
\name{calculate_solution_commonness_site}
\alias{calculate_solution_commonness_site}
\title{Function updates the commonness matrix for site-species solutions for one specific site
Uses equation: Cij = sum(total_species(i) * total_species(j))}
\usage{
calculate_solution_commonness_site(solution_matrix, solution_commonness,
  site)
}
\arguments{
\item{solution_matrix}{predicted species-site matrix}

\item{solution_commonness}{current commonness matrix}

\item{site}{column ID of current site}
}
\value{
matrix of common species between each site
}
\description{
Function updates the commonness matrix for site-species solutions for one specific site
Uses equation: Cij = sum(total_species(i) * total_species(j))
}
