% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_target.R
\name{calculate_target}
\alias{calculate_target}
\title{Function calculate the common species pair-wise matrix (i.e. the target)}
\usage{
calculate_target(alpha, beta)
}
\arguments{
\item{alpha}{predicted alpha diverisity.}

\item{beta}{pairwise matrix of predicted dissimilarity.}
}
\value{
matrix of common species between each site
}
\description{
Function calculate the common species pair-wise matrix (i.e. the target)
}
