% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.check.R
\name{check.packages}
\alias{check.packages}
\title{check.packages}
\usage{
See \url{https://sydneycytometry.org.au/spectre} for usage instructions and vignettes.
}
\value{
returns an error message if one of the common use packages are not installed. Proceeds in order of package importance, and only the first error message encountered will be returned.
}
\description{
This function allows you to check to see if all the common use packages dependencies for Spectre are installed.
}
\examples{
check.packages()

}
\references{
\url{https://sydneycytometry.org.au/spectre}
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
