% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify}
\alias{classify}
\title{classify}
\usage{
classify(x, ...)
}
\arguments{
\item{x}{NO DEFAULT. A dataframe containing cells (rows) vs features/markers (columns) that is the data to be classified.}

\item{type}{DEFAULTS to KNN. Specify the type of classifier to be used. Can be KNN (k-nearest neighbour).}

\item{num.neighbours}{DEFAULTS to 1. When using a k-nearest neighbour classifier, then this parameter specifies the number of nearest neighbours.}

\item{training}{NO DEFAULT. A dataframe containing cells (rows) vs features/markers (columns) that is the data to be use as training.}

\item{training.label}{NO DEFAULT. Name of the column in the training data to be used as the training labels}

\item{new.label}{NO DEFAULT. Name of the new column that contains classification labels.}
}
\value{
...
}
\description{
classify
}
\examples{
classify()

}
\references{
\url{https://sydneycytometry.org.au/spectre}.
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
