% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour.plot.R
\name{colour.plot}
\alias{colour.plot}
\title{colour.plot - Create a dot plot (X vs Y) coloured by a selected column}
\usage{
See \url{https://sydneycytometry.org.au/spectre} for usage instructions and vignettes.
}
\arguments{
\item{d}{data.frame. Input sample. No default.}

\item{x.axis}{Character. Column for X axis. No default.}

\item{y.axis}{Character. Column for Y axis. No default.}

\item{col.axis}{Character. Column for colour. No default.}

\item{col.min.threshold}{Numeric. Define minimum threshold for colour scale. Values below this limit will be coloured as the chosen minimum threshold. Defaults to 0.995.}

\item{col.max.threshold}{Numeric. Define maximum threshold for colour scale. Values above this limit will be coloured as the chosen maximum threshold. Defaults to 0.01.}

\item{title}{Character. Title for the plot. No default.}

\item{colours}{Character. Colour scheme to use. Can be "spectral" or "jet". Defaults to "spectral"}

\item{dot.size}{Numeric. Size of the dots. Defaults to 1.}

\item{align.xy.by}{data.frame. Sample to use to determine minimum and maximum X and Y axis values. No default.}

\item{align.col.by}{data.frame. Sample to use to determine minimum and maximum colour values. No default.}
}
\value{
prints a ggplot.
}
\description{
This function allows you to create a coloured XY plot where each cell is coloured by a selected column. Typically used to plot cells on tSNE1/2 or UMAP1/2 coloured by select cellular markers.
}
\examples{
colour.plot()

}
\references{
\url{https://sydneycytometry.org.au/spectre}
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
