% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonness_complete.R
\name{commonness_complete}
\alias{commonness_complete}
\title{commonness_complete}
\usage{
commonness_complete(species_grid, target)
}
\arguments{
\item{species_grid}{Optimized grid using run_optimization.}

\item{target}{Pairwise matrix of species in common.}
}
\value{
Proportion of correctly solved sites
}
\description{
Counts the number of correctly solved sites
}
\details{

}
\references{
xxx
}
