% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.spectre.R
\name{create.spectre.object}
\alias{create.spectre.object}
\title{Create Spectre object}
\usage{
create.spectre.object(cell_id_col, name = NULL)
}
\arguments{
\item{cell_id_col}{Character. The column in the list of data.tables denoting
the unique identifier of the cells.}

\item{name}{Character vector. DEFAULT = NULL.
Describes the experiment/study associated with the object.}
}
\description{
Create Spectre object
}
