% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.stats.R
\name{create.stats}
\alias{create.stats}
\title{Statistical analysis}
\usage{
create.stats(
  dat,
  use.cols,
  sample.col,
  grp.col,
  comparisons,
  variance.test = "kruskal.test",
  pairwise.test = "wilcox.test",
  corrections = "fdr"
)
}
\description{
Statistical analysis
}
