% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.sumtable.R
\name{create.sumtable}
\alias{create.sumtable}
\title{create.sumtables - create a data.table 'summarising' cellular data by sample and population/cluster.}
\usage{
create.sumtable(dat, sample.col, pop.col, use.cols, annot.cols, counts, func, sep)
}
\arguments{
\item{dat}{NO DEFAULT. A data.table containing cells (rows) vs features/markers (columns). One column must represent sample names, and another must represent populations/clusters.}

\item{sample.col}{NO DEFAULT. Character. Name of the sample column (e.g. "Sample").}

\item{pop.col}{NO DEFAULT. Character. Name of the population/cluster column (e.g. "Population", "Cluster").}

\item{use.cols}{NO DEFAULT. A character vector indicating the columns to be measured (e.g. cellular columns -- c("CD45", "CD3e") etc).}

\item{annot.cols}{DEFAULT = NULL. A character vector indicating the columns to be included as annotation columns (e.g. c("Batch", "Group") etc).}

\item{counts}{DEFAULT = NULL. If you wish to calculate the actual number of cells per sample, a data.table containing the sample names (in column 1) and cell counts per sample (column 2).}

\item{func}{DEFAULT = "median". Can be "median" or "mean". Defines the type of function for calculating MFI data.}

\item{sep}{DEFAULT = " -- ". Character separation of the measurement type and the population (e.g. MFI of CD4 -- T cells)}
}
\description{
This function summarises cellular data and generates a summary data.table
}
\examples{
## Calculate and export results from demonstration data
dat <- Spectre::demo.clustered
counts <- data.frame('Sample' = unique(dat[['Sample']]), 'Counts' = c(rep(100000, 6), rep(1000000, 6)))

sum.dat <- create.sumtable(dat = dat,
                            sample.col = "Sample",
                            pop.col = "Population", 
                            use.cols = names(dat)[c(11:19)],
                            counts = counts)

}
\references{
\url{https://sydneycytometry.org.au/spectre}.
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
