% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{current_solution}
\alias{current_solution}
\title{#' Add information
#'
#' @format matrix
"bci.tree8"
Add information}
\format{
matrix
}
\usage{
current_solution
}
\description{
#' Add information
#'
#' @format matrix
"bci.tree8"
Add information
}
\keyword{datasets}
