% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.asinh.R
\docType{data}
\name{demo.asinh}
\alias{demo.asinh}
\title{demo.asinh - Demo dataset with asinh transformed values}
\format{
A data.table containing arcsinh transformed values.
}
\source{
Thomas M Ashhurst.
}
\usage{
demo.asinh
}
\description{
demo.asinh - Demo dataset with asinh transformed values
}
\examples{
demo.asinh

}
\references{
\url{https://github.com/ImmuneDynamics/Spectre}.
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
\keyword{datasets}
