% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.batches.1.R
\docType{data}
\name{demo.batches.1}
\alias{demo.batches.1}
\title{demo.batches.1 - Samples from two batches with synthetically introduced batch effects for testing.}
\format{
A data.table containing data with batch effects
}
\source{
Thomas M Ashhurst.
}
\usage{
demo.batches.1
}
\description{
demo.batches.1 - Samples from two batches with synthetically introduced batch effects for testing.
}
\examples{
demo.batches.1

}
\references{
\url{https://github.com/ImmuneDynamics/Spectre}.
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
\keyword{datasets}
