% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.batches.R
\docType{data}
\name{demo.batches}
\alias{demo.batches}
\title{demo.batches - Samples from two batches with synthetically introduced batch effects for testing.}
\format{
A data.table containing data with batch effects
}
\source{
Thomas M Ashhurst.
}
\usage{
demo.batches
}
\description{
demo.batches - Samples from two batches with synthetically introduced batch effects for testing.
}
\examples{
demo.batches

}
\references{
\url{https://github.com/ImmuneDynamics/Spectre}.
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
\keyword{datasets}
