% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.clustered.R
\docType{data}
\name{demo.clustered}
\alias{demo.clustered}
\title{demo.clustered - Demo dataset.}
\format{
A dataframe containing bone marrow cells from 6x mock-infected mice and 6x WNV-infected mice, including FlowSOM cluster assignments.
}
\source{
Thomas M Ashhurst.
}
\usage{
demo.clustered
}
\description{
demo.clustered - Demo dataset.
}
\examples{
demo.clustered

}
\references{
\url{https://sydneycytometry.org.au/spectre}.
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
\keyword{datasets}
