% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.dat.R
\docType{data}
\name{demo.dat}
\alias{demo.dat}
\title{demo.dat - Demo dataset.}
\format{A dataframe containing bone marrow cells from 6x mock-infected mice and 6x WNV-infected mice.}
\source{
Thomas M Ashhurst.
}
\usage{
demo.dat
}
\description{
demo.dat - Demo dataset.
}
\examples{
demo.dat

}
\references{
\url{https://sydneycytometry.org.au/spectre}.
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
\keyword{datasets}
