% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.data.R
\docType{data}
\name{demo.data}
\alias{demo.data}
\title{demo.data - Demo data.table with cells (rows) x markers (columns), with clusters, populations, and UMAP values added}
\format{
Demo data.table with cells (rows) x markers (columns), with clusters, populations, and UMAP values added.
}
\source{
Thomas M Ashhurst.
}
\usage{
demo.data
}
\description{
demo.data - Demo data.table with cells (rows) x markers (columns), with clusters, populations, and UMAP values added
}
\examples{
demo.data

}
\references{
\url{https://github.com/ImmuneDynamics/Spectre}.
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
\keyword{datasets}
