% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.exp.R
\docType{demo.exp}
\name{demo.exp}
\alias{demo.exp}
\title{demo.exp - Demo marker expression table with markers (columns) vs populations (rows)}
\format{
Demo marker expression table with markers (columns) vs populations (rows).
}
\source{
Thomas M Ashhurst.
}
\usage{
demo.exp
}
\description{
demo.exp - Demo marker expression table with markers (columns) vs populations (rows)
}
\examples{
demo.exp

}
\references{
\url{https://github.com/ImmuneDynamics/Spectre}.
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
\keyword{datasets}
