% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.spatial.R
\docType{data}
\name{demo.spatial}
\alias{demo.spatial}
\title{demo.spatial - Demo spatial dataset consisting of 6x regions of interest, acquired using Imaging Mass Cytometry  (IMC)}
\format{
An R object (list) containing 6x ROIs, each consisting of raster stacks, masks, and extracted cellular data in a data.table
}
\source{
Thomas M Ashhurst.
}
\usage{
demo.spatial
}
\description{
demo.spatial - Demo spatial dataset consisting of 6x regions of interest, acquired using Imaging Mass Cytometry  (IMC)
}
\examples{
demo.spatial

}
\references{
\url{https://github.com/ImmuneDynamics/Spectre}.
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
\keyword{datasets}
