% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.start.R
\docType{data}
\name{demo.start}
\alias{demo.start}
\title{demo.start - Demo dataset.}
\format{A dataframe containing bone marrow cells from 6x mock-infected mice and 6x WNV-infected mice.}
\source{
Thomas M Ashhurst.
}
\usage{
demo.start
}
\description{
demo.start - Demo dataset.
}
\examples{
demo.start

}
\references{
\url{https://sydneycytometry.org.au/spectre}.
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
\keyword{datasets}
