% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.sum.R
\docType{data}
\name{demo.sum}
\alias{demo.sum}
\title{demo.sum - Demo summary dataset with features/measurments (columns) x samples (rows)}
\format{
Demo summary dataset with features/measurments (columns) x samples (rows).
}
\source{
Thomas M Ashhurst.
}
\usage{
demo.sum
}
\description{
demo.sum - Demo summary dataset with features/measurments (columns) x samples (rows)
}
\examples{
demo.sum

}
\references{
\url{https://sydneycytometry.org.au/spectre}.
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
\keyword{datasets}
