% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.umap.R
\docType{data}
\name{demo.umap}
\alias{demo.umap}
\title{demo.umap - Demo dataset.}
\format{
A dataframe containing bone marrow cells from 6x mock-infected mice and 6x WNV-infected mice including FlowSOM cluster assignments and UMAP coordinates.
}
\source{
Thomas M Ashhurst.
}
\usage{
demo.umap
}
\description{
demo.umap - Demo dataset.
}
\examples{
demo.umap

}
\references{
\url{https://sydneycytometry.org.au/spectre}.
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
\keyword{datasets}
