% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.add.cols.R
\name{do.add.cols}
\alias{do.add.cols}
\title{do.add.cols}
\usage{
do.add.cols(dat, base.col, add.dat, add.by, rmv.ext)
}
\arguments{
\item{dat}{NO DEFAULT. A data.table (or data.frame) containing the data to have new values added to.}

\item{base.col}{NO DEFAULT. Name of the column containing values that new values will be matched to.}

\item{add.dat}{NO DEFAULT. A data table of new values to embed as a new columns, with one column containing values used for matching with the target data.table.}

\item{add.by}{NO DEFAULT. Character, name of the column in add.dat that is used for matching to the taret dataset.}

\item{rmv.ext}{DEFAULTS TO FALSE Logical, can be TRUE or FALSE. Removes a ".csv" or ".fcs" extension from a the 'match.to' vector -- especially useful if 'match.to' is a list of sample names that end in .csv or .fcs.}

\item{mem.ctrl}{DEFAULT = TRUE. Runs gc() (garbage collection) after a number of steps to free up memory that hasn't been released quickly enough.}
}
\value{
Returns the data.table with new columns embedded.
}
\description{
do.add.cols
}
\examples{
add.dt <- data.table('Files' = unique(Spectre::demo.asinh$FileName),
                     'SampleNumber' = c(1:12))

cell.dat <- do.add.cols(dat = Spectre::demo.asinh,
                        base.col = "FileName",
                        add.dat = add.dt,
                        add.by = "Files")

}
\references{
\url{https://sydneycytometry.org.au/spectre}.
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
