% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.add.percell.R
\name{do.add.percell}
\alias{do.add.percell}
\title{do.add.percell}
\usage{
do.add.percell(spatial.dat, percell.dat, roi.col)

do.add.percell(spatial.dat, percell.dat, roi.col, name = "per.cell")
}
\description{
do.add.percell
}
\references{
\url{https://github.com/ImmuneDynamics/Spectre}.
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
