% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.aggregate.R
\name{do.aggregate}
\alias{do.aggregate}
\title{Aggregate data using data.table functions}
\usage{
do.aggregate(dat, use.cols, by, func = "median", dt = TRUE)
}
\description{
Aggregate data using data.table functions
}
