% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.align.R
\name{do.align}
\alias{do.align}
\title{do.align - Function to align multiple batches of a dataset.}
\usage{
do.align(ref.dat, target.dat, batch.col, align.cols, method, goal, 
nQ, Qmin, Qmax, write.ref.fcs, write.target.fcs)
}
\arguments{
\item{ref.dat}{NO DEFAULT. If using method = "CytoNorm", this must be a FlowSOM object created by Spectre::do.prep.fsom. If using "Quantiles", then this must be}

\item{target.dat}{NO DEFAULT. A data.table of data you wish to align}

\item{batch.col}{NO DEFAULT. Character, column that denotes batches}

\item{align.cols}{NO DEFAULT. Character, a vector of columns that you wish to align.}

\item{method}{DEFAULT = "CytoNorm". Character, can be "CytoNorm" or "Quantile".}

\item{goal}{DEFAULT = "Mean". For method = "CytoNorm". Character, the goal to align to. Can either be "mean", or can be the name of one of the batches (e.g. "Batch1").}

\item{nQ}{DEFAULT = 101. For method = "CytoNorm". Numeric, the number of quanitles to use.}

\item{Qmin}{DEFAULT = 0.01. For method = "Quantiles". Numeric, the minimum valus for the lower threshold.}

\item{Qmax}{DEFAULT = 0.99. For method = "Quantiles". Numeric, the maximum value for the upper threshold.}

\item{write.ref.fcs}{DEFAULT = TRUE. Logical, do you want a quick copy of the reference sample FCS files.}

\item{write.target.fcs}{DEFAULT = TRUE. Logical, do you want a quick copy of the ALIGNED target sample FCS files.}

\item{mem.ctrl}{DEFAULT = TRUE. Runs gc() (garbage collection) after a number of steps to free up memory that hasn't been released quickly enough.}
}
\value{
Returns a data.table with selected columns (align.cols) replaced with aligned data.
}
\description{
This function allows you to align multiple batches of a dataset, using either a "Quantile" approach, or using "CytoNorm"
}
\examples{
aligned.dat <- do.align()

}
\references{
Ashhurst, T. M., et al. (2019). \url{https://www.ncbi.nlm.nih.gov/pubmed/31077106}
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
