% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.asinh.R
\name{do.asinh}
\alias{do.asinh}
\title{Transform data using ArcSinh transformation}
\usage{
do.asinh(dat, use.cols, cofactor=5, append.cf=FALSE,
reduce.noise=FALSE)
}
\arguments{
\item{dat}{NO DEFAULT. data.table object. Input data.}

\item{use.cols}{NO DEFAULT. Vector of character column names.
The columns to transform.
These columns will be transformed and added to the data.table as new columns.}

\item{cofactor}{DEFAULT = 5. Co-factor to use for arcsinh transformation.
Can be list of co-factors that matches the columns in \code{use.cols}.
If specifying a list of co-factors make sure each column in \code{use.cols} is assigned
a co-factor (repeat values if necessary).}

\item{append.cf}{DEFAULT = FALSE.
Appends the co-factor used to the end of the name of the transformed columns.}

\item{reduce.noise}{DEFAULT = FALSE.
This is an experimental calculation which should reduce noise
from negative values. Use with caution.}
}
\value{
A data.table with new columns added, that contain asinh
transformed data.
}
\description{
Function to transform data in selected columns using ArcSinh transformation
with a specified co-factor.
}
