% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.calculate.area.R
\name{do.calculate.area}
\alias{do.calculate.area}
\title{do.calculate.area}
\usage{
\method{do}{calculate.area}(dat, region = NULL)
}
\arguments{
\item{dat}{NO DEFAULT. Spatial data list}

\item{region}{DEFAULT = NULL. Name of the region mask, if present.}
}
\description{
do.calculate.area
}
