% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.calculate.earth.mover.dist.R
\name{do.calculate.earth.mover.dist}
\alias{do.calculate.earth.mover.dist}
\title{Calculate earth. mover distance}
\usage{
do.calculate.earth.mover.dist(dat, markers, sample_source_col, batch_id_col)
}
\arguments{
\item{dat}{data.table}

\item{markers}{}

\item{sample_source_col}{}

\item{batch_id_col}{}
}
\description{
Calculate earth. mover distance
}
