% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.clip.R
\name{do.clip}
\alias{do.clip}
\title{Clip data}
\usage{
do.clip(dat, use.cols)

do.clip(dat, use.cols, min.value = NULL, max.value = NULL, append.name = '_clipped')
}
\arguments{
\item{dat}{NO DEFAULT. data.table Input sample.}

\item{use.cols}{NO DEFAULT. Columns to clip}

\item{min.value}{DEFAULT = NULL. Specify the value for which values below will be converted to. If set to NULL, no clipping will occur.}

\item{max.value}{DEFAULT = NULL. Specify the value for which values above will be converted to. If set to NULL, no clipping will occur.}

\item{append.name}{DEFAULT = '_clipped'. Text appended to new column names after clipping.}
}
\value{
A data.table with new columns added, that contain the clipped data.
}
\description{
Clips data using a specified lower and upper value.
E.g. if the upper value is set to 1000, then any values above 1000 will be converted to 1000. No rows are lost, rather the values in those rows are converted.
}
\references{
\url{https://sydneycytometry.org.au/spectre}
}
\seealso{
\url{https://sydneycytometry.org.au/spectre} for usage instructions and vignettes.
}
