% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.clr.R
\name{do.clr}
\alias{do.clr}
\title{Do CLR transformation}
\usage{
do.clr(
  dat,
  data_source,
  output_name,
  use_cols,
  how = c("across_cells", "across_features"),
  verbose = FALSE
)
}
\arguments{
\item{dat}{A spectre object.}

\item{data_source}{Character. The name of the data in Spectre object to
apply CLR transformation to.}

\item{output_name}{Character. What name should the CLR transformed
data be stored under in the Spectre object.}

\item{use_cols}{A vector of character column names to apply CLR transformation to.}

\item{how}{Whether to apply the CLR transformation across cells ("across_cells")
which essentially apply CLR for each feature, or across features ("across_features")
which essentially apply CLR for each cell.}

\item{verbose}{DEFAULT = TRUE.
If TRUE, the function will print progress updates as it executes.}
}
\value{
An updated spectre object
}
\description{
Perform CLR transformation.
}
