% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.combine.cols.R
\name{do.combine.cols}
\alias{do.combine.cols}
\title{Combine two columns}
\usage{
do.combine.cols(dat, col1, col2, na.rm)
}
\arguments{
\item{dat}{NO DEFAULT. Data.table.}

\item{col1}{NO DEFAULT. Name of the column you want to combine into.}

\item{col2}{NO DEFAULT. Name of the column you want to combine from. This column will be deleted after combination into col1.}

\item{na.rm}{DEFAULT = TRUE. If NA values are present in the columns, they will be removed.}
}
\value{
Returns a data.table with adjusted columns
}
\description{
Combine the values of two columns into a single column, with an option to remove any 'NA' values.
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
