% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.convert.to.fold.R
\name{do.convert.to.fold}
\alias{do.convert.to.fold}
\title{do.convert.to.fold}
\usage{
do.convert.to.fold(
  x,
  sample.col,
  group.col,
  ctrl.grp,
  convert.cols,
  log2 = TRUE
)
}
\description{
do.convert.to.fold
}
