% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.create.outlines.R
\name{do.create.outlines}
\alias{do.create.outlines}
\title{do.create.outlines}
\usage{
do.create.outlines(spatial.dat, mask.name, method = "stars")
}
\arguments{
\item{spatial.dat}{NO DEFAULT. Spatial data list}

\item{mask.name}{NO DEFAULT. Name of the mask to create outlines for}

\item{method}{DEFAULT = 'stars'. Can be 'stars' or 'raster'}
}
\description{
do.create.outlines
}
