% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.embed.columns.R
\name{do.embed.columns}
\alias{do.embed.columns}
\title{do.embed.columns}
\usage{
do.embed.columns(dat, base.col, add.dat, add.by, rmv.ext)
}
\arguments{
\item{dat}{NO DEFAULT. A data.table (or data.frame) containing the data to have new values added to.}

\item{base.col}{NO DEFAULT. Name of the column containing values that new values will be matched to.}

\item{add.dat}{NO DEFAULT. A data table of new values to embed as a new columns, with one column containing values used for matching with the target data.table.}

\item{add.by}{NO DEFAULT. Character, name of the column in add.dat that is used for matching to the taret dataset.}

\item{rmv.ext}{DEFAULTS TO TRUE. Logical, can be TRUE or FALSE. Removes the ".csv" or ".fcs" extension from a the 'match.to' vector -- especially useful if 'match.to' is a list of sample names that end in .csv or .fcs.}
}
\value{
Returns the data.table with new columns embedded..
}
\description{
do.embed.columns
}
\examples{

}
\references{
\url{https://sydneycytometry.org.au/spectre}.
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
