% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.extract.cell.dat.R
\name{do.extract.cell.dat}
\alias{do.extract.cell.dat}
\title{do.extract.cell.dat}
\usage{
\method{do}{extract.cell.dat}(dat = spatial.dat, type = "per.cell.means")
}
\description{
do.extract.cell.dat
}
