% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.filter.percell.R
\name{do.filter.percell}
\alias{do.filter.percell}
\title{do.filter.percell}
\usage{
do.filter.percell(spatial.dat, per.cell, to, filter.by)
}
\description{
This is a developmental Spectre-spatial function that is
still in testing phase with limited documentation. We recommend only using
this function if you know what you are doing.
}
