% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.filter.R
\name{do.filter}
\alias{do.filter}
\title{do.filter}
\usage{
do.filter(dat, use.col, values)
}
\arguments{
\item{dat}{NO DEFAULT. A data.table}

\item{use.col}{DEFAULT = NULL. The column to use for re-ordering}

\item{values}{DEFAULT = NULL. A vector of values to use for filtering -- all cells that contain any of the match values will be filtered, and provided in a new data.table. Up to 35 match values can be provided}
}
\description{
This function allows filtering of a data.table using multiple match values -- all cells that contain any of the match values will be filtered, and provided in a new data.table.
}
\references{
\url{https://github.com/ImmuneDynamics/Spectre}.
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
