% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.label.from.polygon.R
\name{do.label.from.polygon}
\alias{do.label.from.polygon}
\title{do.label.from.polygon}
\usage{
do.label.from.polygon(
  spatial.dat,
  cell.dat,
  mask,
  labels,
  name = "Label",
  id.col = "ID",
  roi.col = "ROI"
)
}
\description{
do.label.from.polygon
}
