% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.list.summary.R
\name{do.list.summary}
\alias{do.list.summary}
\title{Summarise the column names and row numbers for elements of a list}
\usage{
do.list.summary(dat)
}
\description{
Summarise the column names and row numbers for elements of a list
}
