% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.logicle.R
\name{do.logicle}
\alias{do.logicle}
\title{Using logicle to transform data}
\usage{
do.logicle(dat, use.cols, linearisation.width, max.scale.val, 
full.transform.width, additional.negative.range)

do.logicle(dat, use.cols, auto.infer.function = TRUE,
linearisation.width = 1.2, max.scale.val = 262144, full.transform.width = 4.5,
additional.negative.range = 0)
}
\arguments{
\item{dat}{NO DEFAULT. data.table. Input data}

\item{use.cols}{NO DEFAULT. Vector of column names to transform}

\item{auto.infer.function}{Default = TRUE.
Automatically infer the logical transformation function based on your data.
If this is set to FALSE, default or overriden
\code{linearisation.width, max.scale.val, full.transform.width, additional.negative.range}
will be used instead to calculate the logical transformation function.}

\item{linearisation.width}{Default = 1.2. Linearisation width in asymptotic decades.
This must be > 0 and determines the slope of transformation at zero.
We found 1.2 works well for flow data, and 0.5 works well for cytof data.
It can be estimated using equation:
\eqn{(m-log10(max.scale.val/|r|))/2}
where r is the most negative value to included in the display.
See \code{max.scale.val} below.}

\item{max.scale.val}{Default = 262144. Maximum scale data value.
It can be 10,000 for common 4 decade data or 262144 (the default value) for 18 bit data range.
This must be greater than 0}

\item{full.transform.width}{Default = 4.5. The full width of the transformed values in asymptotic decades
This must be greater than 0}

\item{additional.negative.range}{Default = 0. Additional negative range to be included in
the transformed value in asymptotic decades.
Value greater than 0 will bring the prescribed additional range into the transformed values}
}
\description{
The function transform data using logicle transformation.
It makes use of flowCore's implementation of logicle transformation, and can
either automatically infer the transformation function based on your data,
or calculate it based on what you want by looking at the default or overriden \code{linearisation.width, max.scale.val, full.transform.width, additional.negative.range}.
Description of some of the parameter is adapted from flowCore's vignette.
For more information on what logicle transformation does, please read the manuscript in the references section.
}
\examples{
library(Spectre)
dat = Spectre::demo.clustered[, 2:9]
use.cols = names(dat)
do.logicle(dat, use.cols = use.cols)

}
\references{
Parks, David R., Mario Roederer, and Wayne A. Moore. "A new “Logicle” display method avoids deceptive effects of logarithmic scaling for low signals and compensated data." Cytometry Part A: The Journal of the International Society for Analytical Cytology 69.6 (2006): 541-551.
}
