% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.mn.R
\name{do.mn}
\alias{do.mn}
\title{do.mn - Multipurpose function for arcsinh transformation, quantile normalisation, noise reduction, and normalisation}
\usage{
do.mn(dat, use.cols)
}
\arguments{
\item{dat}{NO DEFAULT. A data.table}

\item{use.cols}{NO DEFAULT. Vector of column names that you wish to transform}

\item{append.name}{DEFAULT = '_transf'. Text to be appended to resultant column name}

\item{asinh.cf}{DEFAULT = 15. The co-factor to use for arcsinh transformation}

\item{divide.by}{DEFAULT = NULL. Column name to divide data -- where each division is subject to the quantile adjustment, minimum thresholding, and normalisation separately}

\item{Qs}{DEFAULT = NULL. Vector of minimun and maximum quantiles (out of '1') for thresholding (e.g. c(0.005, 0.995)).}

\item{low.threshold}{DEFAULT = NULL. Values below this number will be converted to this number.}

\item{norm}{DEFAULT = c(0,1). New MIN and MAX values for each channel, as c(MIN, MAX).}

\item{info.dir}{DEFAULT = getwd(). Directory address to save a .txt. file with the settings used for the MRN function.}
}
\value{
Returns a data.table with transformed data added as new columns.
}
\description{
This function facilitates multiple data transformation functions in a single process, including arcsinh transformation, noise reduction, and normalisation
}
\examples{
cell.dat <- do.mn()

}
\references{
Ashhurst, T. M., et al. (2019). \url{https://www.ncbi.nlm.nih.gov/pubmed/31077106}
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
