% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.mrn.R
\name{do.mrn}
\alias{do.mrn}
\title{do.mrn - Multipurpose function for arcsinh transformation, Reduction of noise, and Normalisation}
\usage{
do.mrn(dat, use.cols)
}
\arguments{
\item{dat}{NO DEFAULT. A data.table}

\item{use.cols}{NO DEFAULT. Vector of column names that you wish to transform}

\item{asinh.cf}{DEFAULT = 15. The co-factor to use for arcsinh transformation}

\item{low.threshold}{DEFAULT = 0. Values below this number will be converted to this number.}

\item{norm}{DEFAULT = c(0,1). New MIN and MAX values for each channel, as c(MIN, MAX).}
}
\value{
Returns a data.table with transformed data added as new columns.
}
\description{
This function facilitates multiple data transformation functions in a single process, including arcsinh transformation, noise reduction, and normalisation
}
\examples{
cell.dat <- do.mrn()

}
\references{
Ashhurst, T. M., et al. (2019). \url{https://www.ncbi.nlm.nih.gov/pubmed/31077106}
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
