% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.noise.reduce.R
\name{do.noise.reduce}
\alias{do.noise.reduce}
\title{Reduce noise by change values below a specified threshold to that value, reducing the noise generated by 'negative' values}
\usage{
do.noise.reduce(dat, use.cols, cutoffs)
}
\arguments{
\item{dat}{NO DEFAULT. Data table of cells (rows) x features (columns)}

\item{use.cols}{NO DEFAULT. A vector of column names -- specifying which columns will have noise reduction applied to them}

\item{cutoffs}{NO DEFAULT. A vector of numeric cutoff values for each of the use.cols -- values below this threshold will be modified to the threshold value.}
}
\value{
Returns a data.table with the noise-reduced data added as new columns, including a 'noiseRed' suffix
}
\description{
Reduce noise by change values below a specified threshold to that value, reducing the noise generated by 'negative' values
}
\examples{

min(Spectre::demo.asinh$Ly6G_asinh) # Minimum value of CD11b_asinh should be -5.176318

res <- Spectre::do.noise.reduce(dat = Spectre::demo.asinh,
                                use.cols = names(Spectre::demo.asinh)[c(11:19)],
                                cutoffs = rep(0, 9))

}
\references{
\url{https://sydneycytometry.org.au/spectre}.
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
