% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.pull.data.R
\name{do.pull.data}
\alias{do.pull.data}
\title{do.pull.data}
\usage{
do.pull.data(spatial.dat, target.dat)
}
\arguments{
\item{target.dat}{NO DEFAULT. Dataset to pull.}

\item{dat}{NO DEFAULT. Spatial data list.}
}
\description{
do.pull.data
}
