% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.reorder.R
\name{do.reorder}
\alias{do.reorder}
\title{do.reorder - custom row re-ordering of a data.table}
\usage{
do.reorder(dat, use.col, new.order)
}
\arguments{
\item{dat}{NO DEFAULT. A data.table}

\item{use.col}{DEFAULT = NULL. The column to use for re-ordering}

\item{new.order}{DEFAULT = NULL. A vector of row values for the specified column, in the desired order}
}
\description{
This function allows for a user-defined re-ordering of data.table rows
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
Givanna Putri
}
