% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.split.data.R
\name{do.split.data}
\alias{do.split.data}
\title{do.split.data - Function to split a data.tables into multiple files based on time point.}
\usage{
do.split.data(cell.dat. time.point.col)
}
\arguments{
\item{cell.dat}{NO DEFAULT. List of data.tables (or data.frames)}

\item{time.point.col}{NO DEFAULT. Column denoting the time point.}

\item{clustering.cols}{NO DEFAULT. The columns to be used for clustering.}
}
\description{
do.split.data - Function to split a data.tables into multiple files based on time point.
}
\examples{
do.split.data(cell.dat = cell.dat, time.point.col = 'TimePoint', clustering.cols = c("SCA1", "FITC Ly6C"))

}
\references{
\url{https://sydneycytometry.org.au/spectre}.
}
\author{
Givanna Putri, \email{ghar1821@uni.sydney.edu.au}
}
