% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.stats.R
\name{do.stats}
\alias{do.stats}
\title{Statistical analysis}
\usage{
do.stats(
  dat,
  use.cols,
  sample.col,
  grp.col,
  comparisons,
  variance.test = "kruskal.test",
  pairwise.test = "wilcox.test",
  corrections = "fdr"
)
}
\description{
Statistical analysis
}
