% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.subsample.R
\name{do.subsample}
\alias{do.subsample}
\title{Subsample data}
\usage{
do.subsample(dat, method, samp.col, targets, seed)
}
\arguments{
\item{dat}{NO DEFAULT. Input dataframe with cells (rows) vs markers (columns).}

\item{method}{NO DEFAULT. Character. Can be 'random' (downsampling single number from merged data), 'per.sample' (specifying number of cells each sample contributes) or 'min.per.sample' (each sample contributes the same amount of data based on sample with lowest count).}

\item{samp.col}{NO DEFAULT. Character. Name of the column that reflects sample names.}

\item{targets}{NO DEFAULT. List of downsample targets (if random, then absolute target; if per.sample, then targets per sample, must be in the same order as the unique sample names appear in the data frame or data table; can leave blank if min.per.sample).}

\item{seed}{DEFAULT = 42. Numeric. Seed for reproducibility.}
}
\description{
Method to subsample data.
Can subsample by randomly selecting a desired number of cells from all samples (method = "random"), subsample by specifying the exact number of cells for each sample (method= "per.sample"), or by subsampling the same number of cells from each sample based on the sample with the lowest count (method = "min.per.sample").
Useful to decrease total cells for generating dimensionality reduction plots (tSNE/UMAP).
}
\examples{
Spectre::do.subsample(dat = Spectre::demo.start,
                      method = "per.sample",
                      samp.col = "FileName",
                      targets = c(rep(1000,12)))

}
\author{
Thomas Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
Felix Marsh-Wakefield, \email{felix.marsh-wakefield@sydney.edu.au}
}
