% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.zscore.R
\name{do.zscore}
\alias{do.zscore}
\title{do.zscore - Calculate z.score for a dataset}
\usage{
do.zscore(dat, use.cols)
}
\arguments{
\item{dat}{NO DEFAULT. A data.table (or data.frame) containing the data to be converted to z-scores.}

\item{use.cols}{NO DEFAULT. The columns to be used for z-score calculations.}
}
\value{
Returns a new data.table with z-score calculations for each selected column
}
\description{
do.zscore - Calculate z.score for a dataset
}
\examples{

}
