% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.asinh.R
\name{do_actual_transformation}
\alias{do_actual_transformation}
\title{Do actual asinh transformation}
\usage{
do_actual_transformation(
  dat,
  use.cols,
  cofactor = 5,
  cofactor_inference_method = c("flowVS", "top10"),
  append.cf = FALSE,
  reduce.noise = FALSE,
  digits = NULL,
  verbose = TRUE
)
}
\arguments{
\item{dat}{NO DEFAULT.
Either a data.table or a Spectre object to apply arc-sinh transformation to.}

\item{use.cols}{NO DEFAULT.
A vector of character column names to apply arc-sinh transformation to.}

\item{cofactor}{DEFAULT = 5. Co-factor to use for arcsinh transformation.
Can be vector of co-factors that align with columns in 'use.cols'
See details for more information.
It can also be set to NULL to get the function to automatically infer the co-factors.}

\item{cofactor_inference_method}{DEFAULT flowVS.
Whether to automatically calculates co-factor for each marker.
Only used if cofactor is NULL, and will force 'append.cf' to TRUE.
Accepted options: 'flowVS' or 'top10'.
flowVS uses the flowVS::estParamFlowVS (be patient as it takes time).
'top10' uses the 90th percentile of values that are negative (based on absolute value)
as the co-factor, and if values are all positive it will use the 10th percentile as co-factor.}

\item{append.cf}{DEFAULT = FALSE. Appends the co-factor used to the end of
the name of the transformed columns.}

\item{reduce.noise}{DEFAULT = FALSE. This is an experimental calculation
which should reduce noise from negative values. Use with caution.}

\item{digits}{DEFAULT = NULL. Number of decimal places as a limit, not used if NULL.
Values beyond will be rounded.
Equal to the number or less (i.e. if 9 is used, but only 5 digits are present,
then 5 digits will be used). Important to control for small floating point
error differences in different OS.}

\item{verbose}{DEFAULT = TRUE.
If TRUE, the function will print progress updates as it executes.}
}
\value{
a data.table containing asinh transformed markers
}
\description{
Internal function that actually do the asinh transformation
}
