% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.columns.R
\name{embed.columns}
\alias{embed.columns}
\title{embed.columns}
\usage{
embed.columns(x, ...)
}
\value{
Embeds new columns into samples
}
\description{
embed.columns
}
\examples{
embed.columns()

}
\references{
\url{https://sydneycytometry.org.au/spectre}. Helpful examples at \url{https://davetang.org/muse/2018/05/15/making-a-heatmap-in-r-with-the-pheatmap-package/}
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
