% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{estimated_gamma}
\alias{estimated_gamma}
\title{Total (estimated) species in the system.}
\format{
numeric
}
\usage{
estimated_gamma
}
\description{
Total (estimated) species in the system.
}
\keyword{datasets}
