% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.plot.R
\name{factor.plot}
\alias{factor.plot}
\title{factor.plot - Create a factor plot}
\usage{
factor.plot(x, ...)
}
\arguments{
\item{d}{data.frame. Input sample. No default.}

\item{y.axis}{Character. Column for Y axis. No default.}

\item{col.axis}{Character. Column for colour. No default.}

\item{bubble.lab}{Character. Column for bubble labels. Defaults to NULL.}

\item{title}{Character. Title for the plot. No default.}

\item{dot.size}{Numeric. Size of the dots. Defaults to 1.}

\item{align.xy.by}{data.frame. Sample to use to determine minimum and maximum X and Y axis values. No default.}

\item{align.col.by}{data.frame. Sample to use to determine minimum and maximum colour values. No default.

This function allows you to plot cells on an XY plot, coloured by some factor (group, cluster, etc)
factor.plot is a wrapper around the heatmap.2 function from gplots.}

\item{a.axis}{Character. Column for X axis. No default.}
}
\description{
factor.plot - Create a factor plot
}
