% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.colour.plot.R
\name{fast.colour.plot}
\alias{fast.colour.plot}
\title{Create a FAST dot plot (X vs Y) coloured by a selected continuous  (e.g. marker expression) or factorial (e..g. cluster, group) column.}
\usage{
make.colour.plot(dat, x.axis, y.axis, col.axis)
}
\arguments{
\item{dat}{NO DEFAULT. data.table Input sample.}

\item{x.axis}{NO DEFAULT. Character. Column for X axis.}

\item{y.axis}{NO DEFAULT. Character. Column for Y axis.}

\item{col.axis}{DEFAULT = NULL. If not specified, plot is coloured by density. If you provide a character name of a column (e.g. "BV605.Ly6C", "Group", "FlowSOM_metacluster" etc), then each point will be coloured by the value in that column.}

\item{col.type}{DEFAULT = "continuous". Can also be "factor".}

\item{add.label}{DEFAULT = FALSE. Adds labels on the plot at the centroid of each factor. Only works if col.type = "factor".}

\item{hex}{DEFAULT = FALSE. Whether to split the data into bins and show the average expression of the bin. Currently only works when specifying col.axis, so does not work with density plots.}

\item{hex.bins}{DEFAULT = 30. Number of bins to split into. Only used if hex is TRUE.}

\item{colours}{DEFAULT = "spectral". Only used if type = 'colour', ignored if type = 'factor'. Specify a colour scheme. Can be "jet", "spectral", "viridis", "inferno", "magma", or "BuPu". Can also provide a data.table with the first column being the factors, and the second column being the corresponding colour code.}

\item{col.min.threshold}{DEFAULT = 0.01. Numeric. Define minimum threshold for colour scale. Values below this limit will be coloured as the chosen minimum threshold.}

\item{col.max.threshold}{DEFAULT = 0.995 Numeric. Define maximum threshold for colour scale. Values above this limit will be coloured as the chosen maximum threshold.}

\item{align.xy.by}{DEFAULT = dat. data.table Sample to use to determine minimum and maximum X and Y axis values.}

\item{align.col.by}{DEFAULT = dat. data.table. Sample to use to determine minimum and maximum colour values.}

\item{regression.line}{DEFAULT = NULL. Regression line to add to show the trend of the data. This will be passed on to ggplot2::geom_smooth function, and accepts either NULL or a character vector, e.g. "lm", "glm", "gam", "loess" or a function.}

\item{title}{DEFAULT = col.axis. Character. Title for the plot.}

\item{filename}{DEFAULT = NULL. Character. The name of the file to save the plot to.}

\item{dot.size}{DEFAULT = 1. Numeric. Size of the dots.}

\item{plot.width}{DEFAULT = 9. Width of the ggplot when saved to disk.}

\item{plot.height}{DEFAULT = 7. Height of the ggplot when saved to disk.}

\item{nudge_x}{DEFAULT = 0.5. When add.label = TRUE, distance the label is shifted from the centroid point on the X axis.}

\item{nudge_y}{DEFAULT = 0.5. When add.label = TRUE, distance the label is shifted from the centroid point on the Y axis.}

\item{square}{DEFAULT = TRUE. Ensures the plot is saved as a square. Set to FALSE if you want a plot with different X and Y lengths.}

\item{legend.loc}{DEFAULT = 'right'. By default plot legends will be on the right hand side. Can specify the legend location to "bottom" if desired, or 'none' to remove it entirely.}

\item{save.to.disk}{DEFAULT = TRUE. Will save the ggplot to disk. If FALSE, will only show the ggplot.}

\item{path}{DEFAULT = getwd(). The location to save your ggplot. By default, will save to current working directory. Can be overidden.}

\item{blank.axis}{DEFAULT = FALSE Logical, do you want a minimalist graph?}
}
\description{
This function allows you to create a coloured XY plot where each cell is coloured by a
selected column. Typically used to plot cells on tSNE1/2 or UMAP1/2 coloured by select
cellular markers or clusters, samples, groups etc.
}
\examples{
# Load packages
library(Spectre)
package.check()
package.load()

# Read data
cell.dat <- Spectre::demo.umap
cell.dat <- as.data.table(cell.dat)

# Draw plot
Spectre::fast.colour.plot(
  dat = cell.dat,
  x.axis = "UMAP_42_X",
  y.axis = "UMAP_42_Y",
  col.axis = "BV605.Ly6C"
)

}
\references{
\url{https://sydneycytometry.org.au/spectre}
}
\seealso{
\url{https://sydneycytometry.org.au/spectre} for usage instructions and vignettes.
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
Givanna Putri
}
