% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.files.R
\name{file.merge}
\alias{file.merge}
\title{file.merge - ...}
\usage{
file.merge(x, ...)
}
\arguments{
\item{x}{List of data frames}

\item{remove.duplicates}{Do you want to remove duplicates?}
}
\value{
A combined dataframe.
}
\description{
file.merge - ...
}
\examples{
file.merge()

}
\references{
\url{https://sydneycytometry.org.au/spectre}. Helpful examples at \url{https://davetang.org/muse/2018/05/15/making-a-heatmap-in-r-with-the-pheatmap-package/}
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
