% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_solutions_min_conf.R
\name{find_solutions_min_conf}
\alias{find_solutions_min_conf}
\title{Find best solution}
\usage{
find_solutions_min_conf(
  alpha_list,
  total_gamma,
  target,
  max_iterations,
  partial_solution = NULL,
  fixed_species = NULL,
  seed = NA,
  max_repetitions,
  stop_at_optimum = TRUE
)
}
\arguments{
\item{alpha_list}{}

\item{total_gamma}{}

\item{target}{}

\item{max_iterations}{}

\item{partial_solution}{}

\item{fixed_species}{}

\item{seed}{}

\item{max_repetitions}{}

\item{stop_at_optimum}{}
}
\value{

}
\description{
Find best solution
}
