% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_commonness_matrix_from_gdm.R
\name{generate_commonness_matrix_from_gdm}
\alias{generate_commonness_matrix_from_gdm}
\title{generate_commonness_matrix_from_gdm}
\usage{
generate_commonness_matrix_from_gdm(gdm_predictions, n_sites, alpha_list)
}
\value{

}
\description{
Generate a commonness siteXsite matrix from sorensen gdm-package output
}
\details{
details
}
\examples{
example
}
