% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data_raster_rescale.R
\name{generate_data_raster_rescale}
\alias{generate_data_raster_rescale}
\title{generate_data_raster_rescale}
\usage{
generate_data_raster_rescale(landscape, gamma, alpha.min, alpha.max)
}
\arguments{
\item{landscape}{any raster, for example generated with the NLMR package}

\item{gamma}{Total (estimated) species in the system.}

\item{alpha.min}{Minimum number of species in a cell (defined as a fraction of gamma, e.g. 0.1 represents 10\% of gamma)}

\item{alpha.max}{Maximum number of species in a cell (defined as a fraction of gamma, e.g. 0.5 represents 50\% of gamma)}
}
\value{
Rescaled raster landscape
}
\description{
Utility function for raster based data generator to rescale raster values as species richness
}
\details{
This function rescales any raster landscape with numerical values to the desired biodiversity indicators.
It uses the following formula for rescaling:
NewValue = (((OldValue - OldMin) * (NewMax - NewMin)) / (OldMax - OldMin)) + NewMin
}
\examples{
\dontrun{
# 1. Generate random landscape with NLMR:
landscape <- NLMR::nlm_random(ncol=50, nrow=50) 
# 2. Rescale to derive species richness values:
gamma <- 500
alpha <- generate_data_raster_rescale(landscape, gamma, 0.1, 0.4)
landscapetools::show_landscape(alpha)
}
}
