% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data_virtualspecies_to_solution.R
\name{generate_data_virtualspecies_to_solution}
\alias{generate_data_virtualspecies_to_solution}
\title{generate_data_virtualspecies_to_solution}
\usage{
generate_data_virtualspecies_to_solution(spp)
}
\arguments{
\item{species}{raster stack containing presence/absence maps}
}
\value{
matrix (species X sites)
}
\description{
Transforms rasterstack containing presence/absence maps to a solution matrix
}
\details{
The function transforms a raster stack containing presence/absence maps of species to a solution matrix.
This can for example be used for species derive from the virtual species function generate_data_virtualspecies().
The solution matrix can then be used to run the spectre optimization algorithms.
}
\examples{

## This example presents the complete workflow of generating virtual species distributions
## over trasnfromation to solution matrix
## to execution of spectre optimization algorithms:' 


#### Step 1: Generate virtualspecies

# Define parameters:
landscape_size <- 50
gamma <- 100
correlation <- 0 # low values = little spatial autocorrelation
community_size <- 0.1 # 10\% of gamma -> 10 communities
beta <- 0.75  # lower beta = wider species distributions

# Generate species and create list of alpha diversit:
spp <- generate_data_virtualspecies(landscape_size, correlation, community_size, gamma, beta)
alpha <- getValues(sum(spp))

#### Step 2: Transform to solution matrix and calculate target matrix:

solution <- generate_data_virtualspecies_to_solution(spp)
target <- calculate_solution_commonness_rcpp(solution)

#### Step 3: Execute optimization

result <- spectre::run_optimization_min_conf(alpha, gamma, target, max_runs = 20000)

}
