% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.hmap.foldcell.R
\docType{data}
\name{hmap.foldcell}
\alias{hmap.foldcell}
\title{hmap.foldcell - Demo dataset.}
\format{A demo dataframe.}
\source{
Thomas M Ashhurst.
}
\usage{
hmap.foldcell
}
\description{
hmap.foldcell - Demo dataset.
}
\examples{
hmap.foldcell

}
\references{
\url{https://sydneycytometry.org.au/spectre}.
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
\keyword{datasets}
