% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.hmap.mfi.R
\docType{data}
\name{hmap.mfi}
\alias{hmap.mfi}
\title{hmap.mfi - Demo dataset.}
\format{
A demo dataframe.
}
\source{
Thomas M Ashhurst.
}
\usage{
hmap.mfi
}
\description{
hmap.mfi - Demo dataset.
}
\examples{
hmap.mfi

}
\references{
\url{https://sydneycytometry.org.au/spectre}.
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
\keyword{datasets}
