% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.ilastik.prep.R
\name{ilastik.prep}
\alias{ilastik.prep}
\title{ilastik.prep}
\usage{
ilastik.prep(from, to, files)
}
\arguments{
\item{from}{NO DEFAULT. Directory containing FOLDERS (ROIs) or OME.TIFF files}

\item{to}{NO DEFAULT. Directory where you would like to copy selected OME.TIFF files for use in Ilastik}

\item{files}{NO DEFAULT. A list of image names (e.g. c("148Nd_148Nd_CD11b.ome.tiff", "152Sm_152Sm_CD3e.ome.tiff") indicating which images you wish to use in Ilastik.}
}
\value{
...
}
\description{
ilastik.prep
}
\examples{
ilastik.prep()

}
\references{
\url{https://sydneycytometry.org.au/spectre}.
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
