% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.asinh.R
\name{infer_cofactor}
\alias{infer_cofactor}
\title{Internal function which infer the co-factor using flowVS or top10 method.}
\usage{
infer_cofactor(value, cofactor_inference_method, use.cols)
}
\arguments{
\item{value}{a data.table containing only the markers to be transformed.}

\item{cofactor_inference_method}{which inferrence method to use? flowVS or top10.}

\item{use.cols}{A vector of character column names to apply arc-sinh transformation to.}
}
\description{
Internal function which infer the co-factor using flowVS or top10 method.
}
\author{
Felix Marsh-Wakefield
}
