% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled.factor.plot.R
\name{labelled.factor.plot}
\alias{labelled.factor.plot}
\title{labelled.factor.plot - Create a factor plot}
\usage{
labelled.factor.plot(x, ...)
}
\arguments{
\item{d}{data.frame. Input sample. No default.}

\item{y.axis}{Character. Column for Y axis. No default.}

\item{col.axis}{Character. Column for colour. No default.}

\item{title}{Character. Title for the plot. Defaults to the name of the colour axis.}

\item{dot.size}{Numeric. Size of the dots. Defaults to 1.}

\item{align.xy.by}{data.frame. Sample to use to determine minimum and maximum X and Y axis values. No default.}

\item{align.col.by}{data.frame. Sample to use to determine minimum and maximum colour values. No default.

This function allows you to plot cells on an XY plot, coloured by some factor (group, cluster, etc) with labels included.
factor.plot is a wrapper around the heatmap.2 function from gplots.}

\item{a.axis}{Character. Column for X axis. No default.}
}
\description{
labelled.factor.plot - Create a factor plot
}
