% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.packages.R
\name{load.packages}
\alias{load.packages}
\title{load.packages}
\usage{
See \url{https://sydneycytometry.org.au/spectre} for usage instructions and vignettes.
}
\value{
loads all the common use package libraries.
}
\description{
This function allows you to load all of the common use packages dependencies for Spectre.
}
\examples{
load.packages()

}
\references{
\url{https://sydneycytometry.org.au/spectre}
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
