% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rasters.R
\name{load_raster}
\alias{load_raster}
\title{Function to help find and load required raster files}
\usage{
load_raster(path, pattern)
}
\arguments{
\item{path}{The directory path to the required data.}

\item{pattern}{The type of file the rasters are stored as.}
}
\value{
Raster stack of imported files.
}
\description{
Function to help find and load required raster files
}
