% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.bubble.plot.R
\name{make.bubble.plot}
\alias{make.bubble.plot}
\title{Make Bubble Plot}
\usage{
make.bubble.plot(dt, markers_to_plot, cluster, col.scheme = "RdYlBu")
}
\arguments{
\item{dt}{Data.table object containing the data to plot}

\item{markers_to_plot}{Vector of markers to plot}

\item{cluster}{Name of the column in dt which specify the clusters}

\item{col.scheme}{Colour scheme to use. Must be something from
the RColorBrewer or the viridis package.
Default to 'RdYlBu' in the RColorBrewer package}
}
\value{
ggplot object containing the plot
}
\description{
Dot (or bubble) plot which visualises the marker expression changes and the
percentage of cells captured across different clusters.
Experimental feature.
}
\author{
Givanna Putri
}
