% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.colour.plot.R
\name{make.colour.plot}
\alias{make.colour.plot}
\title{make.colour.plot - Create a dot plot (X vs Y) coloured by a selected continuous column (e.g. marker expression)}
\usage{
make.colour.plot(dat, x.axis, y.axis, col.axis)

See \url{https://sydneycytometry.org.au/spectre} for usage instructions and vignettes.
}
\arguments{
\item{dat}{NO DEFAULT. data.table Input sample.}

\item{x.axis}{NO DEFAULT. Character. Column for X axis.}

\item{y.axis}{NO DEFAULT. Character. Column for Y axis.}

\item{col.axis}{NO DEFAULT. Character. Column for colour.}

\item{title}{DEFAULT = col.axis. Character. Title for the plot.}

\item{col.min.threshold}{DEFAULT = 0.01. Numeric. Define minimum threshold for colour scale. Values below this limit will be coloured as the chosen minimum threshold.}

\item{col.max.threshold}{DEFAULT = 0.995 Numeric. Define maximum threshold for colour scale. Values above this limit will be coloured as the chosen maximum threshold.}

\item{colours}{DEFAULT = "spectral". Character. Colour scheme to use. Can be "spectral", "jet", "viridis", "magma", or "inferno".}

\item{dot.size}{DEFAULT = 1. Numeric. Size of the dots.}

\item{align.xy.by}{DEFAULT = dat. data.table Sample to use to determine minimum and maximum X and Y axis values.}

\item{align.col.by}{DEFAULT = dat. data.table. Sample to use to determine minimum and maximum colour values.}

\item{save.to.disk}{DEFAULT = TRUE. Will save the ggplot to disk. If FALSE, will only show the ggplot.}

\item{path}{DEFAULT = getwd(). The location to save your ggplot. By default, will save to current working directory. Can be overidden.}

\item{plot.width}{DEFAULT = 9. Width of the ggplot when saved to disk.}

\item{plot.height}{DEFAULT = 7. Height of the ggplot when saved to disk.}

\item{blank.axis}{DEFAULT = FALSE Logical, do you want a minimalist graph?}
}
\value{
prints and saves a ggplot.
}
\description{
This function allows you to create a coloured XY plot where each cell is coloured by a selected column. Typically used to plot cells on tSNE1/2 or UMAP1/2 coloured by select cellular markers.
}
\examples{
make.colour.plot(dat = demo.umap, x.axis = "UMAP_X", y.axis = "UMAP_Y", col.axis = "BV605.Ly6C")

}
\references{
\url{https://sydneycytometry.org.au/spectre}
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
