% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.density.plot.R
\name{make.density.plot}
\alias{make.density.plot}
\title{Make a density plot (using ggpointdensity::geom_pointdensity())}
\usage{
make.multi.marker.plot(dat, x.axis, y.axis, title, colours, dot.size, align.xy.by, align.col.by, save.to.disk, path, plot.width,  plot.height, blank.axis)
}
\arguments{
\item{dat}{NO DEFAULT. data.frame.}

\item{x.axis}{NO DEFAULT. Character. Column for X axis.}

\item{y.axis}{NO DEFAULT. Character. Column for Y axis.}

\item{title}{DEFAULT = "Density". Character. Title for the plot.}

\item{colours}{DEFAULT = "viridis". Character. Colour scheme to use. Can be "spectral", "jet", "viridis", "magma", or "inferno".}

\item{dot.size}{DEFAULT = 1. Numeric. Size of the dots.}

\item{align.xy.by}{DEFAULT = dat. data.table Sample to use to determine minimum and maximum X and Y axis values.}

\item{align.col.by}{DEFAULT = dat. data.table. Sample to use to determine minimum and maximum colour values.}

\item{save.to.disk}{DEFAULT = TRUE. Will save the ggplot to disk. If FALSE, will only show the ggplot.}

\item{path}{DEFAULT = getwd(). The location to save your ggplot. By default, will save to current working directory. Can be overidden.}

\item{plot.width}{DEFAULT = 9. Width of the ggplot when saved to disk.}

\item{plot.height}{DEFAULT = 7. Height of the ggplot when saved to disk.}

\item{blank.axis}{DEFAULT = FALSE. Logical, do you want a minimalist graph?}
}
\description{
Method to create a density plot.
A density plot can summarise the cellular distribution across a given plot.
Can be used to plot any two markers.
Uses "ggplot2" for plots, "ggpointdensity" for adding density to plots, "RColorBrewer" for colour schemes.
}
\examples{
# Create density plot on demonstration dataset
Spectre::make.density.plot(dat = Spectre::demo.umap,
                  x.axis = "UMAP_42_X",
                  y.axis = "UMAP_42_Y",
                  save.to.disk = FALSE)

}
\author{
Felix Marsh-Wakefield, \email{felix.marsh-wakefield@sydney.edu.au}
}
