% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.factor.plot.R
\name{make.factor.plot}
\alias{make.factor.plot}
\title{make.factor.plot - Create a dot plot (X vs Y) coloured by a selected factor column (e.g. clusters, samples)}
\usage{
make.factor.plot(dat, x.axis, y.axis, col.axis)

See \url{https://sydneycytometry.org.au/spectre} for usage instructions and vignettes.
}
\arguments{
\item{dat}{NO DEFAULT. data.table Input sample.}

\item{x.axis}{NO DEFAULT. Character. Column for X axis.}

\item{y.axis}{NO DEFAULT. Character. Column for Y axis.}

\item{col.axis}{NO DEFAULT. Character. Column for colour, should be a factor (e.g. cluster, group, sample, etc).}

\item{add.label}{DEFAULT = FALSE. Add's a label for the entries in 'col.axis' onto the plot -- useful for labelling clusters (e.g. 1, 2, 3 etc) or populations (T cell, B cell, etc).}

\item{nudge_x}{DEFAULT = 0.5. Distance to nudge the label from the point (on the x axis) when add.label = TRUE.}

\item{nudge_y}{DEFAULT = 0.5. Distance to nudge the label from the point (on the y axis) when add.label = TRUE.}

\item{title}{DEFAULT = col.axis. Character. Title for the plot.}

\item{dot.size}{DEFAULT = 1. Numeric. Size of the dots.}

\item{align.xy.by}{DEFAULT = dat. data.table Sample to use to determine minimum and maximum X and Y axis values.}

\item{align.col.by}{DEFAULT = dat. data.table. Sample to use to determine minimum and maximum colour values.}

\item{blank.axis}{DEFAULT = FALSE. Logical, do you want a minimalist graph?}

\item{col.min.threshold}{DEFAULT = 0.01. Numeric. Define minimum threshold for colour scale. Values below this limit will be coloured as the chosen minimum threshold.}

\item{col.max.threshold}{DEFAULT = 0.995 Numeric. Define maximum threshold for colour scale. Values above this limit will be coloured as the chosen maximum threshold.}
}
\value{
prints and saves a ggplot.
}
\description{
This function allows you to create a coloured XY plot where each cell is coloured by a selected column. Typically used to plot cells on tSNE1/2 or UMAP1/2 coloured by select factor, such as clusters, samples, groups, batches etc.
}
\examples{
make.factor.plot(dat = demo.umap, x.axis = "UMAP_X", y.axis = "UMAP_Y", col.axis = "FlowSOM_metacluster")

}
\references{
\url{https://sydneycytometry.org.au/spectre}
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
