% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.heatmap.R
\name{make.heatmap}
\alias{make.heatmap}
\title{make.heatmap - Create a heatmap}
\usage{
make.heatmap(x, ...)
}
\arguments{
\item{x}{data.frame. Clusters/populations vs markers (MFI) or Clusters/populations vs samples or (MFI or cell numbers). No default.}

\item{sample.col}{Character. Specify the name of the column that indicates samples. No default.}

\item{annot.cols}{Character. Columns which contain values that you do NOT want to plot in the heatmap, e.g. sample name, group name, Live/Dead etc. No default.}

\item{plot.title}{Character.}

\item{transpose}{Logical. Do you want to transpose the heatmap. Defaults to FALSE.}

\item{is.fold}{Logical. TRUE for fold-change heatmap, FALSE for normal values heatmap. Defaults to FALSE.}

\item{fold.max.range}{Numeric. For fold-change heatmaps, what is the maxium colour value that should be plotted. Defaults to 3.}

\item{fold.min.range}{Numeric. For fold-change heatmaps, what is the minimum colour value that should be plotted. Defaults to -3.}

\item{normalise}{Logical. Only applies to standard heatmaps (i.e. when is.fold = FALSE). TRUE to normalise each column between 0 and 1, FALSE to plot the raw values. Defaults to TRUE.}

\item{dendrograms}{Character. Do you want to include dendrograms on columns/rows. Can be "both", "row", "column", or "none. Defaults to "both.}

\item{n.row.groups}{Numeric.}

\item{n.col.groups}{Numeric.}

\item{row.sep}{Numeric.}

\item{col.sep}{Numeric.}

\item{y.margin}{Numeric.}

\item{x.margin}{Numeric.}

\item{standard.colours}{Character.}

\item{fold.colours}{Character.}

\item{colour.scheme}{Character.}
}
\value{
prints a heatmap with dendrograms.
}
\description{
This function allows you to create a coloured heatmap for clusters vs markers (coloured by MFI) or clusters vs samples (coloured by MFI or cell counts). Can be set to plot 'fold-change' values that are provided in log2.
make.heatmap is a wrapper around the heatmap.2 function from gplots.
}
\examples{
make.heatmap()

}
\references{
\url{https://sydneycytometry.org.au/spectre}
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
