% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.mds.plot.R
\name{make.mds.plot}
\alias{make.mds.plot}
\title{Draw MDS plot}
\usage{
make.mds.plot(
  dat,
  sample_col,
  markers,
  colour_by,
  add_point_label = TRUE,
  font_size = 4
)
}
\arguments{
\item{dat}{}

\item{sample_col}{}

\item{markers}{}

\item{colour_by}{}
}
\description{
Experimental feature. Still very much work in progress.
}
