% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.multi.marker.plot.R
\name{make.multi.marker.plot}
\alias{make.multi.marker.plot}
\title{Make mulitple plots for each marker}
\usage{
make.multi.marker.plot(dat, x.axis, y.axis, plot.by, density.plot, align.xy.by, align.col.by, colours, figure.title, dot.size, col.min.threshold, col.max.threshold, path, plot.width,  plot.height, blank.axis, save.each.plot)
}
\arguments{
\item{dat}{NO DEFAULT. A data frame containing all the data you wish to plot}

\item{x.axis}{NO DEFAULT. X axis}

\item{y.axis}{NO DEFAULT. Y axis}

\item{plot.by}{NO DEFAULT. A list or vector of column names (markers) to plot.}

\item{density.plot}{DEFAULT = FALSE. Logical. Creates a density plot (using 'ggpointdensity' package).}

\item{align.xy.by}{NO DEFAULT. Align X and Y to a dataset}

\item{align.col.by}{NO DEFAULT. Align colour to a dataset}

\item{colours}{DEFAULTS to 'spectral'. What colour scheme do you want to use. Only used if type = 'colour', ignored if type = 'factor'. Can be 'jet', 'spectral', 'viridis', 'inferno', or 'magma'.}

\item{figure.title}{DEFAULTS to 'Marker expression'. Figure title}

\item{dot.size}{DEFAULTS to 1. Size of dots}

\item{col.min.threshold}{DEFAULTS to 0.01. Minimum threshold for colour scale.}

\item{col.max.threshold}{DEFAULTS to 0.995. Maximum threshold for colour scale.}

\item{path}{DEFAULTS TO getwd() -- i.e. the current working directory. Path to the desired output directory}

\item{plot.width}{DEFAULTS to 9.}

\item{plot.height}{DEFAULTS to 7.}

\item{blank.axis}{DEFAULT = FALSE. Logical, do you want a minimalist graph?}

\item{save.each.plot}{DEFAULT = FALSE. Do you want to save each plot?}
}
\description{
Method to create multiple plots for each marker.
This function allows you to create a grid of plots, where the cells are subsetted by a certain factor (e.g. one sample per plot).
These can then be coloured by a marker or by another factor (e.g Group).
Makes use of Spectre functions make.colour.plot and make.density.plot.
}
\examples{
# Create grid of plots on demonstration data
Spectre::make.multi.marker.plot(dat = Spectre::demo.umap,
                                x.axis <- "UMAP_42_X",
                                y.axis <- "UMAP_42_Y",
                                plot.by <- c("AF700.CD45", "APCCy7.CD48", "BV605.Ly6C", "DL800.SA.Bio.Ly6G"),
                                )

}
\author{
Thomas Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
Felix Marsh-Wakefield, \email{felix.marsh-wakefield@sydney.edu.au}
}
