% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.multi.plot.R
\name{make.multi.plot}
\alias{make.multi.plot}
\title{Make multiple plots for multiple columns and/or multiple sample/group/clusters}
\usage{
make.multi.plot(dat, x.axis, y.axis, plot.by)
}
\arguments{
\item{dat}{NO DEFAULT. A data frame containing all the data you wish to plot}

\item{x.axis}{NO DEFAULT. X axis}

\item{y.axis}{NO DEFAULT. Y axis}

\item{plot.by}{NO DEFAULT. A vector of character names for the columns you wish to plot.}

\item{divide.by}{DEFAULT = NULL. Here you can specify a character name of a column you wish to use to divide up the dataset.}

\item{add.density}{DEFAULT = FALSE. Can specify to add a density plot at the end the series of colour plots}

\item{col.type}{DEFAULT = "continuous". Can also be "factor".}

\item{figure.title}{DEFAULT = "Multi plot". Also used as the prefix for the saved file name.}

\item{align.xy.by}{NO DEFAULT. Align X and Y to a dataset}

\item{align.col.by}{NO DEFAULT. Align colour to a dataset}

\item{colours}{DEFAULTS to 'spectral'. What colour scheme do you want to use. Only used if type = 'colour', ignored if type = 'factor'. Can be 'jet', 'spectral', 'viridis', 'inferno', 'magma', or "BuPu".}

\item{dot.size}{DEFAULT = 1. Numeric. Size of the dots.}

\item{col.min.threshold}{DEFAULT = 0.01. Numeric. Define minimum threshold for colour scale. Values below this limit will be coloured as the chosen minimum threshold.}

\item{col.max.threshold}{DEFAULT = 0.995 Numeric. Define maximum threshold for colour scale. Values above this limit will be coloured as the chosen maximum threshold.}

\item{path}{DEFAULTS TO getwd() -- i.e. the current working directory. Path to the desired output directory}

\item{plot.width}{DEFAULTS to 9.}

\item{plot.height}{DEFAULTS to 7.}

\item{blank.axis}{DEFAULT = FALSE. Logical, do you want a minimalist graph?}

\item{save.each.plot}{DEFAULT = FALSE. Do you want to save each plot?}
}
\description{
Method to create multiple plots for each marker.
This function allows you to create a grid of plots, where the cells are plotted by a series of columns, and/or subsetted by a certain factor (e.g. one sample per plot).
Makes use of Spectre functions make.colour.plot and make.density.plot.
}
\examples{
# Create grid of plots on demonstration data
Spectre::make.multi.plot(dat = as.data.table(Spectre::demo.umap),
                          x.axis = "UMAP_42_X",
                          y.axis = "UMAP_42_Y",
                          plot.by = c("BV605.Ly6C", "BUV737.B220", "AF700.CD45"))

}
\author{
Thomas Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
Felix Marsh-Wakefield, \email{felix.marsh-wakefield@sydney.edu.au}
}
