% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.network.plot.R
\name{make.network.plot}
\alias{make.network.plot}
\title{Draw network diagram}
\usage{
make.network.plot(dat, timepoint.col, timepoints, cluster.col, marker.cols,
node.size = 13, arrow.length = 3, arrow.head.gap = 4, standard.colours = 'Spectral')
}
\arguments{
\item{dat}{NO DEFAULT. Data.frame. Data to be clustered.}

\item{timepoint.col}{NO DEFAULT. Column name which represents the time point of each cell (data point) in dat.}

\item{timepoints}{NO DEFAULT. The time points (in order).}

\item{cluster.col}{NO DEFAULT. Column denoting the cluster id.}

\item{marker.cols}{NO DEFAULT. Vector of column names denoting the markers to plot.}

\item{node.size}{DEFAULT = 'auto'. By default, set the size of the node to proportion of cells in the cluster. Set this to specific number if you want the node size to be the same regardless.}

\item{min.node.size}{DEFAULT = 6. Minimum node size if node size is scaled to proportion of cells in cluster. Otherwise not used.}

\item{max.node.size}{DEFAULT = 12. Maximum node size if node size is scaled to proportion of cells in cluster. Otherwise not used.}

\item{arrow.length}{DEFAULT = 3. Length of the arrow connecting nodes.}

\item{arrow.head.gap}{DEFAULT = 4. The gap between head of the arrow and node.}

\item{standard.colours}{DEFAULT = "Spectral". Colour scheme for the markers. Spectral or Inferno or Viridis.}
}
\description{
Method to draw a network diagram.
This rely on a clustering performed by ChronoClust.
}
\author{
Givanna Putri, \email{givanna.haryonoputri@sydney.edu.au}
}
