% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.spatial.plot.R
\name{make.spatial.plot}
\alias{make.spatial.plot}
\title{make.spatial.plot}
\usage{
make.spatial.plot(
  spatial.dat,
  image.roi,
  image.channel,
  mask.outlines = NULL,
  cell.dat = NULL,
  cell.col = NULL,
  image.y.flip = TRUE,
  image.mask.size = 0.1,
  image.mask.colour = "gold",
  image.min.threshold = 0,
  image.max.threshold = 0.99,
  image.blank = FALSE,
  cell.x = "x",
  cell.y = "y",
  cell.col.type = "numeric",
  cell.colours = "spectral",
  cell.col.min.threshold = 0.01,
  cell.col.max.threshold = 0.995,
  title = paste0(image.roi),
  dot.size = 1,
  dot.alpha = 1,
  align.xy.by = cell.dat,
  align.col.by = cell.dat,
  save.to.disk = TRUE,
  path = getwd(),
  plot.width = 9,
  plot.height = 7,
  blank.axis = FALSE
)
}
\arguments{
\item{spatial.dat}{NO DEFAULT.A spatial data list}

\item{image.roi}{NO DEFAULT.The name of the ROI to plot}

\item{image.channel}{NO DEFAULT.Name of the channel to plot}

\item{mask.outlines}{DEFAULT = NULL. Name of the mask for outlines}

\item{cell.dat}{DEFAULT = NULL. Character name for the cellular dataset in the spatial data list, or a data.table object.}

\item{cell.col}{DEFAULT = NULL. Name of cellular dataset to colour points by}

\item{image.y.flip}{DEFAULT = TRUE. Flips the Y-axis orientation.}

\item{image.mask.size}{DEFAULT = 0.1}

\item{image.mask.colour}{DEFAULT = "gold"}

\item{image.min.threshold}{DEFAULT = 0.00. Lower threshold for plot signal. Values below this target will be clipped.}

\item{image.max.threshold}{DEFAULT = 0.99. Upper threshold for plot signal. Values above this target will be clipped.}

\item{image.blank}{DEFAULT = FALSE. Will blank the plot image.}

\item{cell.x}{DEFAULT = "x". Column name for the 'x' coordinate in data.table}

\item{cell.y}{DEFAULT = "y". Column name for the 'y' coordinate in data.table}

\item{cell.col.type}{DEFAULT= "numeric". Can be 'factor'.}

\item{cell.colours}{= DEFAULT = "spectral".}

\item{cell.col.min.threshold}{= DEFAULT = 0.01. Values below this target will be clipped.}

\item{cell.col.max.threshold}{= DEFAULT = 0.995. Values above this target will be clipped.}

\item{title}{DEFAULT = paste0(image.roi)}

\item{dot.size}{DEFAULT = 1}

\item{dot.alpha}{DEFAULT = 1}

\item{align.xy.by}{DEFAULT = cell.dat}

\item{align.col.by}{DEFAULT = cell.dat}

\item{save.to.disk}{DEFAULT = TRUE}

\item{path}{DEFAULT = getwd(). Path for saving image files}

\item{plot.width}{DEFAULT = 9}

\item{plot.height}{DEFAULT = 7}

\item{blank.axis}{DEFAULT = FALSE}
}
\description{
make.spatial.plot
}
