% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.sumtable.R
\name{make.sumtable}
\alias{make.sumtable}
\title{make.sumtable}
\usage{
make.sumtable(x, ...)
}
\arguments{
\item{x}{NO DEFAULT. A dataframe containing cells (rows) vs features/markers (columns). One column must represent sample names, and another must represent populations/clusters.}

\item{sample.col}{NO DEFAULT.}

\item{clust.col}{NO DEFAULT.}

\item{annot.col.nums}{DEFAULTS TO NULL.}

\item{do.frequencies}{DEFAULTS TO TRUE.}

\item{do.exp.per.sample}{DEFAULTS TO TRUE.}

\item{do.exp.per.marker}{DEFAULTS TO TRUE.}

\item{cells.per.tissue}{DEFAULTS TO NULL.}

\item{fun.type}{DEFAULTS TO "median".}

\item{do.foldchange}{DEFAULTS TO FALSE.}

\item{group.col}{DEFAULTS TO NULL.}

\item{control.group}{DEFAULTS TO NULL.}
}
\value{
...
}
\description{
make.sumtable
}
\examples{
make.sumtable()

}
\references{
\url{https://sydneycytometry.org.au/spectre}.
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
