% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{minimal_example_data}
\alias{minimal_example_data}
\title{List with example data created using the \code{gdm} package}
\format{
list
}
\usage{
minimal_example_data
}
\description{
List with example data created using the \code{gdm} package
}
\keyword{datasets}
