% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.plots.R
\name{multi.plot}
\alias{multi.plot}
\title{multi.plot}
\usage{
multi.plot(x, ...)
}
\arguments{
\item{d}{NO DEFAULT. A data frame containing all the data you wish to plot}

\item{type}{DEFAULTS to 'factor' to create factor plots. Will add 'colour' option to create colour plots in next update.}

\item{x.axis}{X axis}

\item{y.axis}{Y axis}

\item{plot.by}{What factor should the plots be divided by}

\item{align.xy.by}{Align X and Y to a dataset}

\item{align.col.by}{Align colour to a dataset}

\item{colour}{What feature should the plot be coloured by}

\item{figure.title}{Figure title}

\item{dot.size}{Size of dots}
}
\description{
multi.plot
}
