% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{normalise.data}
\alias{normalise.data}
\title{normalise.data - ...}
\usage{
normalise.data(x, ...)
}
\arguments{
\item{data}{NO DEFAULT. A dataframe containing cells (rows) vs features/markers (columns) to be normalised.

Normalise the data to range between 0 and 1.}
}
\value{
Normalised data
}
\description{
normalise.data - ...
}
