% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.check.R
\name{package.check}
\alias{package.check}
\title{package.check - a function to check the installation of all required packages.}
\usage{
See \url{https://sydneycytometry.org.au/spectre} for usage instructions and vignettes.
}
\value{
returns an error message if one of the common use packages are not installed. Proceeds in order of package importance, and only the first error message encountered will be returned.
}
\description{
This function allows you to check to see if all the common use packages dependencies for Spectre are installed.
}
\examples{
package.check()

}
\references{
\url{https://sydneycytometry.org.au/spectre}
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
