% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.install.R
\name{package.install}
\alias{package.install}
\title{package.install - a function to install packages required for Spectre.}
\usage{
package.install()
}
\arguments{
\item{type}{DEFAULT = "general". If "general", then checks for the packages required for general Spectre usage. If "spatial", then checks for additional packages required for spatial analysis.}

\item{update}{DEFAULT = FALSE. If FALSE, will only install packages that are not already installed -- no updates of packages will be performed. If TRUE, will install and update packages.}
}
\value{
returns an error message if one of the common use packages are not installed. Proceeds in order of package importance, and only the first error message encountered will be returned.
}
\description{
This function allows you to check to see if all the common use packages dependencies for Spectre are installed. Will only install if the package has not been installed, will not update packages.
}
\examples{
package.install()
package.install(type = "spatial")

}
\references{
\url{https://sydneycytometry.org.au/spectre}
}
\seealso{
See \url{https://sydneycytometry.org.au/spectre} for usage instructions and vignettes.
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
