% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.load.R
\name{package.load}
\alias{package.load}
\title{package.load - a function to load (library) all required packages.}
\usage{
See \url{https://sydneycytometry.org.au/spectre} for usage instructions and vignettes.
}
\arguments{
\item{type}{DEFAULT = "general". If "general", then loads packages required for general Spectre usage. If "spatial", then loads  additional packages required for spatial analysis. If "ML", then loads additional packages required for machine-learing functionality.}
}
\value{
loads all the common use package libraries.
}
\description{
This function allows you to load all of the common use packages dependencies for Spectre.
}
\examples{
package.load()

}
\references{
\url{https://sydneycytometry.org.au/spectre}
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
