% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_energy.R
\name{plot_energy}
\alias{plot_energy}
\title{plot_energy}
\usage{
plot_energy(species_grid)
}
\arguments{
\item{species_grid}{Optimized grid using run_optimization.}
}
\value{
ggplot
}
\description{
Plot energy
}
\details{
Plot energy over time
}
\references{
xxx
}
