% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_error.R
\name{plot_error}
\alias{plot_error}
\title{plot_error}
\usage{
plot_error(x)
}
\arguments{
\item{x}{Results object from run_optimization_min_conf}
}
\value{
ggplot
}
\description{
Plot the absolute error
}
\details{
Plot error over time
}
\references{
xxx
}
