% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_pair_dissimilarity.R
\name{predict_pair_dissimilarity}
\alias{predict_pair_dissimilarity}
\title{Function creates a predicted pairwise dissimilarity for every pair of sites in the landscape}
\usage{
predict_pair_dissimilarity(x, y, envdat, gdmmodel)
}
\arguments{
\item{x}{x-coord.}

\item{y}{y-coord.}

\item{envdat}{A dataframe of environmental variable values for each landuse.}

\item{gdmmodel}{GDM model.}
}
\value{
A pairwise matrix of predicted dissimilarity.
}
\description{
Function creates a predicted pairwise dissimilarity for every pair of sites in the landscape
}
