% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.align.R
\name{prep.align}
\alias{prep.align}
\title{prep.align - Prepare reference data into a FlowSOM object}
\usage{
prep.align()
}
\arguments{
\item{dat}{NO DEFAULT. A data.table consisting of the 'refernece' data you will use to train the alignment algorithm}

\item{cellular.cols}{NO DEFAULT. A vector of column names from the data.table that contain 'cellular' markers}

\item{cluster.cols}{NO DEFAULT. A vector of column names from the data.table that contain markers you wish to use for clusteirng}

\item{batch.col}{NO DEFAULT. Name of the column that contains batch names}

\item{sample.col}{DEFAULT = NULL. Name of the column that contains sample names}

\item{dir}{DEFAULT = getwd(). Sets the working directory to operate from. Because this function involves some reading/writing of files, it's best to set this to somewhere static in case the active working directory moves to a subfolder, and then doesn't return because the function runs into an error.}

\item{xdim}{DEFAULT = 5. Size of X-axis of FlowSOM grid.}

\item{ydim}{DEFAULT = 5. Size of Y-axis of FlowSOM grid.}

\item{meta.k}{DEFAULT = 10. Number of metaclusters. If set to 1, will map all cells to a single metacluster}

\item{seed}{DEFAULT = 42. Seed for reproducibility.}

\item{mem.ctrl}{DEFAULT = TRUE. Allows the function to clear held memory on occasion.}
}
\value{
Returns an object which represents the alignment model. In this preparation stage, it contains the FlowSOM object containing the reference data. The 'train.align' function can then be used to calculate the conversions between batches for each metacluser.
}
\description{
This function allows you to prepare reference data ahead of performing batch alignment.
}
\examples{
align.model <- prep.align()

}
\references{
Ashhurst, T. M., et al. (2019). \url{https://www.ncbi.nlm.nih.gov/pubmed/31077106}
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
