% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.optim_sol.R
\name{print.optim_sol}
\alias{print.optim_sol}
\title{print.optim_sol}
\usage{
\method{print}{optim_sol}(x, ...)
}
\arguments{
\item{x}{optim_sol object with randomized patterns.}

\item{...}{Arguments passed to \code{cat}.}
}
\value{
void
}
\description{
Print method for optim_sol object
}
\details{
Printing method for optimized estimate of community composition.
}
\seealso{
\code{\link{run_optimization_min_conf}}
}
