% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.files.R
\name{read.files}
\alias{read.files}
\title{read.files - Function to read data from CSV or FCS files into a list}
\usage{
read.files(file.loc = getwd(), file.type = ".csv", do.embed.file.names = TRUE)
}
\arguments{
\item{file.loc}{What is the location of your files? Defaults to current working directory (getwd()).}

\item{file.type}{What type of files do you want to read. Can be ".csv" or ".fcs". Defaults to ".csv".}

\item{do.embed.file.names}{Do you want to embed each row (cell) of each file with the name name? Defaults to TRUE.}
}
\value{
a list (called data.list) of dataframes -- one file per dataframe.
}
\description{
This function allows you to read in sample files (.csv or .fcs) into a list called 'data.list'.
}
\examples{
read.files()

}
\references{
Ashhurst, T. M., et al. (2019). \url{https://www.ncbi.nlm.nih.gov/pubmed/31077106}
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
