% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.spatial.files.R
\name{read.spatial.files}
\alias{read.spatial.files}
\title{Read TIFF files into R and and create spatial data object}
\usage{
read.spatial.files()
}
\arguments{
\item{rois}{vector of ROI names (directory names)}

\item{roi.loc}{working directory for ROIs}

\item{multi.tiff}{Default FALSE. Currently we don't support reading in multipage TIFF files}

\item{correct.extent}{Correct extent so that the minimum is 0,0. Default TRUE}

\item{flip.y}{Flip the data arrangement for the y-axis. Default TRUE}

\item{value.modifier}{Data modifier based on image processing. Default 65535}

\item{ext}{Default = ".tif" (which will recognise '.tiff' files as well)}
}
\value{
Returns a spatial data object.
}
\description{
Read TIFF files into R and and create spatial data object
}
\examples{

}
