% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.align.R
\name{run.align}
\alias{run.align}
\title{run.align - Run the alignment model on a target data.table}
\usage{
run.align()
}
\arguments{
\item{dat}{NO DEFAULT. A data.table consisting of the 'refernece' data you will use to train the alignment algorithm}

\item{model}{NO DEFAULT. A batch alignment conversion model object created by the prep.align() and train.align() functions.}

\item{batch.col}{NO DEFAULT. Column name of the data.table that contains batch labels}

\item{append.name}{DEFAULT = "_aligned". This will be appended to the column names containing the aligned data}

\item{dir}{DEFAULT = getwd(). Sets the working directory to operate from. Because this function involves some reading/writing of files, it's best to set this to somewhere static in case the active working directory moves to a subfolder, and then doesn't return because the function runs into an error.}

\item{mem.ctrl}{DEFAULT = TRUE. Allows the function to clear held memory on occasion.}
}
\value{
Returns a data.table with aligned data added in new columns.
}
\description{
This function allows you to prepare reference data ahead of performing batch alignment.
}
\examples{
cell.dat <- run.align()

}
\references{
Ashhurst, T. M., et al. (2019). \url{https://www.ncbi.nlm.nih.gov/pubmed/31077106}
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
