% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.chronoclust.R
\name{run.chronoclust}
\alias{run.chronoclust}
\title{run.chronoclust - ...}
\usage{
run.chronoclust(x, ...)
}
\arguments{
\item{config.xml.file.location}{NO DEFAULT. Location of xml file describing the parameter values for Chronoclust to operate on.}

\item{input.xml.file.location}{NO DEFAULT. Location of xml file describing the location of the dataset (for each timepoint) for Chronoclust to operate on.}

\item{output.dir.location}{NO DEFAULT. The directory for Chronoclust to write its result to.}

\item{virtualenv.name}{DEFAULTS to r-chronoclust. The name of the virtual environment to run Chronoclust on.}

\item{create.virtualenv}{DEFAULTS to TRUE. Whether to create the virtual environment to run Chronoclust on.}

\item{delete.virtualenv}{DEFAULTS to TRUE. Whether to delete the virtual environment to run Chronoclust on after it finishes.

This function run a time-series clustering and cluster tracking algorithm Chronoclust.
Chronoclust is originally written in python. Hence it requires Reticulate to run.
Reticulate executes Chronoclust on a virtual environment.}
}
\description{
run.chronoclust - ...
}
