% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{run.classifier}
\alias{run.classifier}
\title{run.classifier - ...}
\usage{
run.classifier(x, ...)
}
\arguments{
\item{train.data}{NO DEFAULT. A dataframe containing cells (rows) vs features/markers (columns) to be used to train classifier. 
Only include columns/features/markers to be used to train the classifier i.e. the ones that are useful in identifying cells (which population it belongs to).}

\item{train.label}{NO DEFAULT. A character vector containing the label (population name) for each cell in the train.data. Must be "CHARACTER" vector.}

\item{unlabelled.data}{NO DEFAULT. A dataframe containing cells (rows) vs features/markers (columns) that is the data classified.}

\item{num.neighbours}{DEFAULTS to 1. When using a k-nearest neighbour classifier, then this parameter specifies the number of nearest neighbours.

Train a k-NN classifier on a training data, and use it to classify an unlabelled data. 
Note that for the classifier to work as intended, the unlabelled.data has to be normalised to the range of the train.data.
Note make sure that train.data and unlabelled.data has exactly the same features/markers.}
}
\value{
The predicted label for the unlabelled data.
}
\description{
run.classifier - ...
}
