% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.fast.umap.R
\name{run.fast.umap}
\alias{run.fast.umap}
\title{DEPRECATED Faster version of run umap.}
\usage{
run.fast.umap(
  dat,
  use.cols,
  umap.x.name = "UMAP_X",
  umap.y.name = "UMAP_Y",
  umap.seed = 42,
  n_threads = detectCores() - 1,
  n_sgd_threads = "auto",
  batch = TRUE,
  ...
)
}
\arguments{
\item{dat}{Data.table containing your cytometry data.}

\item{use.cols}{Markers to be used to calculate UMAP stored in a vector.}

\item{umap.x.name}{DEFAULT = "UMAP_X". Character. Name of UMAP x-axis.}

\item{umap.y.name}{DEFAULT = "UMAP_Y". Character. Name of UMAP y-axis.}

\item{umap.seed}{DEFAULT = 42. Numeric. Seed value for reproducibility.}

\item{n_threads}{DEFAULT \code{detectCores()-1}. Numeric. Number of threads to use (except during stochastic gradient descent). For nearest neighbor search, only applies if \code{nn_method = "annoy"}. If \code{n_threads > 1}, then the Annoy index will be temporarily written to disk in the location determined by tempfile.}

\item{n_sgd_threads}{DEFAULT "auto". Number of threads to use during stochastic gradient descent. If set to > 1, then be aware that if \code{batch = FALSE}, results will not be reproducible, even if \code{set.seed} is called with a fixed seed before running. Set to "auto" to use the same value as n_threads.}

\item{batch}{DEFAULT TRUE. If set to TRUE, then embedding coordinates are updated at the end of each epoch rather than during the epoch. In batch mode, results are reproducible with a fixed random seed even with n_sgd_threads > 1, at the cost of a slightly higher memory use. You may also have to modify learning_rate and increase n_epochs, so whether this provides a speed increase over the single-threaded optimization is likely to be dataset and hardware-dependent.}

\item{...}{Parameters to be passed to uwot::umap. See their vignette for more info.}
}
\description{
15/6/2023: This function has been merged into run.umap.
To use, please run run.umap with parameter fast set to TRUE.
}
\details{
This umap function runs the implementation provided by the uwot package.
\emph{Note, by default, the calculation will be run in parallel.}
To be more specific, \code{n_threads} and \code{n_sgd_threads} will be automatically set to maximum number
of cores in your computer - 1.
However, you can override this by setting both to NULL to run it in serial.
See uwot::umap vignette for more information.
This umap implementation is \emph{\strong{much much}} faster than the one provided by the
umap package.
Once this has been long established, we will deprecate the old one.

As of Version 0.1.11, it is now possible to get reproducible results (for a given value of set.seed) when running the optimization step with multiple threads (\code{n_sgd_threads} greater than 1).
You may need to increase n_epochs to get similar levels of convergence.
To run in this mode, set batch = TRUE.
}
\author{
Givanna Putri
}
