% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.minmax.normaliser.R
\name{run.minmax.normaliser}
\alias{run.minmax.normaliser}
\title{run.minmax.normaliser - ...}
\usage{
run.minmax.normaliser(x, ...)
}
\arguments{
\item{data}{NO DEFAULT. A dataframe containing cells (rows) vs features/markers (columns) to be normalised.

Normalise the data to range between 0 and 1.}
}
\value{
Normalised data
}
\description{
run.minmax.normaliser - ...
}
