% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.phenograph.R
\name{run.phenograph}
\alias{run.phenograph}
\title{run.phenograph - Run phenograph clustering}
\usage{
run.phenograph()
}
\arguments{
\item{ref.dat}{NO DEFAULT. A data.table consisting of the 'refernece' data you will use to train the alignment algorithm}
}
\value{
Returns a data.table with clustering results added
}
\description{
This function allows you to perform phenograph clustering on a data.table
}
\examples{
\dontrun{
cell.dat <- Spectre::demo.clustered
cell.dat <- Spectre::do.subsample(cell.dat, 5000)

use.cols <- c("NK11_asinh", "CD3_asinh", "CD45_asinh", 
"Ly6G_asinh", "CD11b_asinh", "B220_asinh", "CD8a_asinh", 
"Ly6C_asinh", "CD4_asinh")

cell.dat <- run.phenograph(dat = cell.dat, use.cols = use.cols)
}
}
\references{
Ashhurst, T. M., et al. (2019). \url{https://www.ncbi.nlm.nih.gov/pubmed/31077106}
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
