% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.prepare.chronoclust.R
\name{run.prepare.chronoclust}
\alias{run.prepare.chronoclust}
\title{run.prepare.chronoclust - This function will prepare a Python environment to run ChronoClust.
You need to install anaconda package management software for this to work.
If Reticulate is kind, it might prompt you to install it when running this function if you haven't got it installed.
However, please do not count on it.
If it does not prompt you, and you encounter an error such as no anaconda installation is found,
please visit https://www.anaconda.com/products/individual to manually download and install anaconda.}
\usage{
run.prepare.chronoclust(environment_name, environment_path, create_environment, install_dependencies)
}
\arguments{
\item{environment_name}{NO DEFAULT. Character. The name of the anaconda environment where chronoclust will execute.}

\item{environment_path}{DEFAULT NULL. Character. For expert users only. If you have custom
anaconda installation i.e. not installed to default location, you can specify
the location here.}

\item{create_environment}{DEFAULT TRUE. Logical. Whether to newly create the environment or not. If FALSE, make sure the
environment exists in your anaconda installation.}

\item{install_dependencies}{DEFAULT TRUE. Logical. Whether to install chronoclust and all its dependencies into the environment.
If TRUE, it will install all the dependencies, and restart R session.
When R session is restarted, the variables you have previously assigned
will remain intact, but you will need to reload all the libraries again.}
}
\description{
For advanced users, if your anaconda is installed in custom location,
you can specify this location as environment_path parameter.
}
\examples{
run.prepare.chronoclust(environment_name = "chronoclust-R", create_environment = TRUE, install_dependencies = TRUE)

}
\references{
\url{https://sydneycytometry.org.au/spectre}.
}
\author{
Givanna Putri, \email{ghar1821@uni.sydney.edu.au}
}
