% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.rpca.label.transfer.R
\name{run.rpca.label.transfer}
\alias{run.rpca.label.transfer}
\title{Run rPCA for label transfer}
\usage{
run.rpca.label.transfer(
  dat,
  cytometry_data_source,
  citeseq_data_source,
  output_name,
  celltype_label_col,
  use_cols = NULL,
  k_anchor = 20,
  seed = 42,
  verbose = TRUE
)
}
\arguments{
\item{dat}{NO DEFAULT.
A Spectre object containing the cytometry and CITESeq data to
perform label transfer.}

\item{cytometry_data_source}{NO DEFAULT. Character. The name of the cytometry
data in dat.}

\item{citeseq_data_source}{NO DEFAULT. Character. The name of the CITEseq
data in dat.}

\item{output_name}{NO DEFAULT. Character. What name do you want to store
the resulting predicted label under in dat?}

\item{celltype_label_col}{NO DEFAULT. The column in the CITEseq data which
denotes the cell type/label each cell in the CITEseq data.}

\item{use_cols}{DEFAULT = NULL.
A vector of marker/adt names to use to perform label transfer.
If left as the default value NULL, the function will automatically infer
markers/adts that are common between the CITEseq and cytometry data.
See details for more information.}

\item{k_anchor}{DEFAULT 20. Passed to Seurat's FindIntegrationAnchors function.
Essentially, it determines the number of neighbors (k) to use when
Seurat's FindIntegrationAnchors is picking anchors.}

\item{seed}{DEFAULT 42. Seed used when running PCA.}

\item{verbose}{DEFAULT = TRUE.
If TRUE, the function will print progress updates as it executes.}
}
\value{
Spectre object with new element stored in output_name
}
\description{
Use Seurat rPCA to perform label transfer from a CITEseq data to a cytometry data.
}
\details{
\subsection{Missing markers in use_cols}{

If you supply markers name in use_cols, and it turns out that some of them are
not available in either the CITEseq data, the cytometry data, or both, these
markers will be abandoned.
}

\subsection{No common markers/adts were found}{

Regardless of whether you supply some markers in use_cols or you get the function
to infer them, if as it turns out there are no markers that are common between the
CITEseq data and the cytometry data (and use_cols if you supply it with values),
the function will stop running.
}
}
