% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.spatial.analysis.R
\name{run.spatial.analysis}
\alias{run.spatial.analysis}
\title{run.spatial.analysis}
\usage{
run.spatial.analysis(
  dat,
  sample.col,
  pop.col,
  annot.cols = NULL,
  region.col = NULL,
  area.table = NULL,
  adj.dist = 100,
  x.col = "x",
  y.col = "y",
  distribution = TRUE,
  composition = TRUE,
  counts = TRUE,
  counts.per.area = TRUE,
  distance = TRUE,
  adjacency = TRUE,
  func = "mean"
)
}
\arguments{
\item{dat}{NO DEFAULT. Data.table}

\item{sample.col}{NO DEFAULT. Column that denotes 'samples'}

\item{pop.col}{NO DEFAULT. Column that denotes 'samples'}

\item{annot.cols}{DEFAULT = NULL. Annotation columns.}

\item{region.col}{DEFAULT = NULL. Create a 'total' by default, and add specific ones if requested here}

\item{area.table}{DEFAULT = NULL. Calculate 'total' automatically}

\item{adj.dist}{DEFAULT = 100}

\item{x.col}{DEFAULT = 'x'}

\item{y.col}{DEFAULT = 'y'}

\item{distribution}{DEFAULT = TRUE}

\item{composition}{DEFAULT = TRUE}

\item{counts}{DEFAULT = TRUE}

\item{counts.per.area}{DEFAULT = TRUE}

\item{distance}{DEFAULT = TRUE}

\item{adjacency}{DEFAULT = TRUE}

\item{func}{DEFAULT = 'mean'}
}
\description{
run.spatial.analysis
}
