% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.tsne.R
\name{run.tsne}
\alias{run.tsne}
\title{run.tsne - ...}
\usage{
run.tsne(x, use.cols, ...)
}
\arguments{
\item{x}{data.frame. No default.}

\item{use.cols}{Vector of numbers, reflecting the columns to use for clustering. No default.}

\item{tsne.seed}{Numeric. Seed value for reproducibility. Default = 42.}

\item{dims}{Number of dimensions for output results, either 2 or 3. Default = 2.}

\item{initial_dims}{Number of dimensions retained in initial PCA step. Default = 50.}

\item{perplexity}{Default = 30.}

\item{theta}{Use 0.5 for Barnes-Hut tSNE, 0.0 for exact tSNE (takes longer). Default = 0.5.}

\item{check_duplicates}{Default = FALSE.}

\item{pca}{Runs PCA prior to tSNE run. Default = TRUE.}

\item{max_iter}{Maximum number of iterations. Default = 1000.}

\item{verbose}{Default = TRUE.}

\item{is_distance}{Experimental, using X as a distance matrix. Default = FALSE.}

\item{Y_init}{Recommend NULL for random initialisation. Default = NULL.}

\item{stop_lying_iter}{Number of iterations of early exaggeration. Default = 250.}

\item{mom_switch_iter}{Number of iterations before increased momentum of spread. Default = 250.}

\item{momentum}{Initial momentum of spread. Default = 0.5.}

\item{final_momentum}{Momentum of spread at 'final_momentum'. Default = 0.8.}

\item{eta}{Learning rate. Default = 200.}

\item{exaggeration_factor}{Factor used during early exaggeration. Default = 12.0.

This function runs tSNE on a dataframe with cells (rows) vs markers (columns), and returns 'res' with result columns. Uses the Rtsne package. For more information on parameter choices and effects, check out https://distill.pub/2016/misread-tsne/.}
}
\description{
run.tsne - ...
}
