% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.umap.v2.R
\name{run.umap.v2}
\alias{run.umap.v2}
\title{V2 version of run umap. Experimental. Use at your own risk.}
\usage{
run.umap.v2(
  dat,
  use.cols,
  umap.x.name = "UMAP_X",
  umap.y.name = "UMAP_Y",
  umap.seed = 42,
  ...
)
}
\arguments{
\item{...}{Parameters to be passed to uwot::umap.
See their vignette for more info.}
}
\description{
V2 version of run umap. Experimental. Use at your own risk.
}
