% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.umap.R
\name{run_actual_umap}
\alias{run_actual_umap}
\title{Internal function which actually run umap.}
\usage{
run_actual_umap(
  dat,
  use.cols,
  umap.x.name = "UMAP_X",
  umap.y.name = "UMAP_Y",
  umap.seed = 42,
  neighbours = 15,
  n_components = 2,
  metric = "euclidean",
  n_epochs = 200,
  input = "data",
  init = "spectral",
  min_dist = 0.1,
  set_op_mix_ratio = 1,
  local_connectivity = 1,
  bandwidth = 1,
  alpha = 1,
  gamma = 1,
  negative_sample_rate = 5,
  a_gradient = NA,
  b_gradient = NA,
  spread = 1,
  transform_state = 42,
  knn.repeats = 1,
  verbose = TRUE,
  umap_learn_args = NA,
  fast = FALSE,
  n_threads = "auto",
  n_sgd_threads = "auto",
  batch = TRUE
)
}
\description{
Internal function which actually run umap.
}
