% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_optimization.R
\name{run_optimization}
\alias{run_optimization}
\title{run_optimization}
\usage{
run_optimization(alpha_list, total_gamma, target, cycles, required_D,
  patience)
}
\arguments{
\item{alpha_list}{Matrix of predicted alpha diversity in each cell.}

\item{total_gamma}{Total (estimated) species in the system.}

\item{target}{pairwise matrix of species in common.}

\item{cycles}{number of loops before stopping.}

\item{required_D}{maximum D required to stop.}

\item{patience}{number of loops with no improvement before stopping.}
}
\value{
Best matrix of common species between each site.
}
\description{
xxx
}
\details{
This is the function which runs the optimization algorithm.
There are a number of steps involved
\enumerate{
\item generate initial solution site x species matrix (only needs to be run once per landscape)
\item alter solution site x species matrix and select best site x site matrix based on Dsol
\item Run optimization algorithm until some stop requirement is met
}
}
\references{
xxx
}
