% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_optimization_backtracking.R
\name{run_optimization_backtracking}
\alias{run_optimization_backtracking}
\title{run_optimization}
\usage{
run_optimization_backtracking(
  alpha_list,
  total_gamma,
  target,
  max_runs,
  energy_threshold,
  seed = 0,
  verbose = TRUE
)
}
\arguments{
\item{alpha_list}{Matrix of predicted alpha diversity in each cell.}

\item{total_gamma}{Total (estimated) species in the system.}

\item{target}{Pairwise matrix of species in common.}

\item{max_runs}{Max number of loops before stopping.}

\item{energy_threshold}{Optimization stops if energy threshold is reached. This is set as a value between 0 and 1 determining the proportion of error accepted}

\item{seed}{Seed for random number generator. seed = 0 means that a time stamp is used as seed.}

\item{verbose}{It TRUE, progress report is printed}
}
\value{
Best matrix of common species between each site.
}
\description{
xxx
}
\details{
This is the function which runs the optimization algorithm.
}
\references{
xxx
}
