% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_optimization_min_conf_1.R
\name{run_optimization_min_conf_1}
\alias{run_optimization_min_conf_1}
\title{run_optimization_min_conf_1}
\usage{
run_optimization_min_conf_1(
  alpha_list,
  total_gamma,
  target,
  fixed_species = NULL,
  partial_solution = NULL,
  max_iterations,
  energy_threshold,
  seed = 0,
  verbose = TRUE
)
}
\arguments{
\item{alpha_list}{Matrix of predicted alpha diversity in each cell.}

\item{total_gamma}{Total (estimated) species in the system.}

\item{target}{Pairwise matrix of species in common.}

\item{fixed_species}{Fixed partial solution with species that are considered as given. Those species are not going to be changed during optimization.}

\item{partial_solution}{Partial or complete initial solution as a start for the optimization.}

\item{max_iterations}{Max number of loops before stopping.}

\item{energy_threshold}{Optimization stops if energy threshold is reached. This is set as a value between 0 and 1 determining the proportion of error accepted}

\item{seed}{Seed for random number generator. seed = 0 means that a time stamp is used as seed.}

\item{verbose}{It TRUE, progress report is printed}
}
\value{
Best matrix of common species between each site.
}
\description{
xxx
}
\details{
This is the function which runs the optimization algorithm.
}
\references{
xxx
}
