% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{show,Spectre-method}
\alias{show,Spectre-method}
\title{Print SpectreObject}
\usage{
\S4method{show}{Spectre}(object)
}
\arguments{
\item{SpectreObject}{A SpectreObject}
}
\description{
Print SpectreObject
}
