% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectre.object.R
\name{show,SpectreObject-method}
\alias{show,SpectreObject-method}
\title{Print SpectreObject}
\usage{
\S4method{show}{SpectreObject}(object)
}
\arguments{
\item{SpectreObject}{A SpectreObject}
}
\description{
Print SpectreObject
}
