% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_environmental_variables.R
\name{simulate_environmental_data}
\alias{simulate_environmental_data}
\title{creates a dataframe of simulated environmental variable values for each landuse.
This is done by simple sampling from the data.}
\usage{
simulate_environmental_data(lutdf, env)
}
\arguments{
\item{lutdf}{A table of site locations and landuses.}

\item{env}{A table of environmental variable values linked to each land use.}
}
\value{
A dataframe of simulated environmental variable values for each landuse.
}
\description{
creates a dataframe of simulated environmental variable values for each landuse.
This is done by simple sampling from the data.
}
