% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Spectre-class}
\alias{Spectre-class}
\title{Spectre Class}
\description{
Spectre class is an extension the R native list class designed to store and
manage cytometry data and the results of analysis associated with the data.
It is specifically developed to store them as a list of data.tables.
The types of data and analysis results that can be stored in this list includes
(but not limited to) raw data, asinh-transformed data, and results from downstream
analyses such as clustering or UMAP.
See details for more information.
}
\details{
Each element within the list must be a data.table.
All of these elements \emph{must} include a column which uniquely identify each cell.
The name of this column must be the same for all data.tables in the list and
must be speified in the cell_id_col slot.
Elements in the list should be named descriptively (e.g., "cyto" for raw data, "cyto_asinh" for asinh-transformed data)
but custom naming is supported to accommodate different preferences (e.g., "cyto_raw" for raw data).
}
\section{Slots}{

\describe{
\item{\code{cell_id_col}}{Character. Specifies the name of the column that uniquely identifies each cell within the data tables.
This column must be present in all the data.tables in the list.}

\item{\code{name}}{Character vector. Describes the experiment/study associated with the object.}

\item{\code{metadata}}{List. Stores the metadata associated with each dataset
stored in the object.
One element per dataset.}

\item{\code{spatial}}{Reserved for future use to store spatial data.}

\item{\code{other}}{List. For storing any additional data that does not fit within the list
or other defined slots.}
}}

