% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectre-package.R
\docType{package}
\name{Spectre-package}
\alias{Spectre}
\alias{Spectre-package}
\title{Spectre: High-dimensional cytometry and imaging analysis.}
\description{
A computational toolkit in R for the integration, exploration, and analysis of high-dimensional single-cell cytometry and imaging data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ImmuneDynamics/Spectre}
  \item Report bugs at \url{https://github.com/ImmuneDynamics/Spectre/issues}
}

}
\author{
\strong{Maintainer}: Thomas Ashhurst \email{thomas.ashhurst@sydney.edu.au}

Authors:
\itemize{
  \item Felix Marsh-Wakefield \email{felix.marsh-wakefield@sydney.edu.au}
  \item Givanna Putri \email{putri.g@wehi.edu.au}
}

}
\keyword{internal}
