% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectre-package.R
\docType{package}
\name{spectre}
\alias{spectre}
\alias{spectre-package}
\title{spectre}
\description{
lalalalaalal
}
\author{
\strong{Maintainer}: Craig Simpkins \email{simpkinscraig063@gmail.com}

Authors:
\itemize{
  \item Sebastian Hanss
  \item Maximilian Hesselbarth
  \item Matthias Spangenberg
  \item Jan Salecker
}

}
