% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/do.asinh.R
\docType{class}
\name{SpectreObject-class}
\alias{SpectreObject-class}
\alias{SpectreObject}
\alias{do.asinh,SpectreObject-method}
\title{SpectreObject Class}
\usage{
SpectreObject(cytometry_data, citeseq_data = data.table())

\S4method{do.asinh}{SpectreObject}(
  dat,
  use.cols,
  slot_name = NULL,
  cofactor = 5,
  append.cf = FALSE,
  reduce.noise = FALSE,
  digits = NULL,
  verbose = TRUE
)
}
\description{
SpectreObject Class
}
\details{
You should use this object if you want to jointly analyse cytometry and
CITEseq data.

If you only want to analyse cytometry data, read the file in as data.table
and run Spectre functions as per normal.
There is no need to use this object.

Both \code{cytometry_data} and \code{citeseq_data} slots must be filled.
By default, if you don't pass any data.table into both slots, they will
both be empty.
If you specify something for \code{cytometry_data} you MUST specify a data.table
for \code{citeseq_data}.
}
\section{Slots}{

\describe{
\item{\code{cytometry_data}}{A data.table slot storing cytometry data.
Rows are cells, columns are markers/features.}

\item{\code{citeseq_data}}{A data.table slot storing CITEseq data.
Rows are cells, columns are antibodies.}
}}

\examples{
# Create an instance of SpectreObject with empty slots

}
