% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.subsample.R
\name{subsample}
\alias{subsample}
\title{subsample - ...}
\usage{
subsample(x, ...)
}
\arguments{
\item{samp.col}{Character. Name of the column that reflects sample names.}

\item{targets}{List of downsample targets (if random, then absolute target; if per.sample, then targets per sample, must be in the same order as the unique sample names appear in the data frame or data table). No default.}

\item{seed}{Numeric. Seed for reproducibility. No default.

This function facilitates downsampling of a dataframe.}

\item{x.}{Input dataframe with cells (rows) vs markers (columns). No default.}

\item{method.}{Character. Can be 'random' or 'per.sample. No default}
}
\description{
subsample - ...
}
