% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumtable-percent.pos.R
\name{sumtable.percent.pos}
\alias{sumtable.percent.pos}
\title{sumtable.percent.pos}
\usage{
sumtable.percent.pos(x, sample.name, cluster.name, Markers, Cutoff)
}
\description{
sumtable.percent.pos
}
