% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{target_matrix}
\alias{target_matrix}
\title{Pairwise matrix of species in common.}
\format{
matrix
}
\usage{
target_matrix
}
\description{
Pairwise matrix of species in common.
}
\keyword{datasets}
