% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.cytonorm.R
\name{train.cytonorm}
\alias{train.cytonorm}
\title{train.cytonorm - Prepare reference data into a FlowSOM object}
\usage{
train.align()
}
\arguments{
\item{method}{DEFAULT = 'cytonorm'. In future additional methods for batch alignment will be added.}

\item{cytonorm.goal}{DEFAULT = 'mean'. Target values for alignment. Can be 'mean' for the average of all batches, or a specific batch can be selected.}

\item{cytonorm.nQ}{DEFAULT = 101. Number of quantiles.}

\item{dir}{DEFAULT = getwd(). Sets the working directory to operate from. Because this function involves some reading/writing of files, it's best to set this to somewhere static in case the active working directory moves to a subfolder, and then doesn't return because the function runs into an error.}

\item{mem.ctrl}{DEFAULT = TRUE. Allows the function to clear held memory on occasion.}

\item{dat}{NO DEFAULT. A data.table consisting of the 'refernece' data you will use to train the alignment algorithm}

\item{cellular.cols}{NO DEFAULT. A vector of column names from the data.table that contain the markers to be aligned}
}
\value{
Returns the alignment model object, with an additional 'conversions' element added.
}
\description{
This function allows you to learn the conversions to use for batch alignment.
}
\references{
Ashhurst, T. M., et al. (2019). \url{https://www.ncbi.nlm.nih.gov/pubmed/31077106}
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
