% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.knn.classifier.R
\name{train.knn.classifier}
\alias{train.knn.classifier}
\title{train.knn.classifier - ...}
\usage{
train.knn.classifier(x, ...)
}
\arguments{
\item{train.data}{NO DEFAULT. A dataframe containing cells (rows) vs features/markers (columns) to be used to train classifier. 
Only include columns/features/markers to be used to train the classifier i.e. the ones that are useful in identifying cells (which population it belongs to).}

\item{label}{NO DEFAULT. A character vector containing the label (population name) for each cell in the train.data. Must be "CHARACTER" vector.}

\item{min.num.neighbours}{DEFAULTS to 1. When using a k-nearest neighbour (KNN) classifier, this parameter specifies the minimum number of nearest neighbours used to train the KNN classifier.}

\item{max.num.neighbours}{DEFAULTS to 1. When using a k-nearest neighbour (KNN) classifier, this parameter specifies the maximum number of nearest neighbours used to train the KNN classifier.}

\item{evaluation.metric}{DEFAULTS to 'Accuracy'. How do you want the classifier performance to be evaluated? By measuring accuracy (pass "Accuracy") or Cohen Kappa score (pass "Kappa").}

\item{num.folds}{DEFAULTS to 10. Number of chunks the training data is to be split into. The classifier will then take 1 chunk for testing the classifier (after it's trained), and use the remaining chunk to train the classifier.}

\item{num.repeats}{DEFAULTS to 10. Number of time the classifier will be trained (per number of neighbours). For each repeat, different chunk will be used for testing and training.

Train a k-nearest neighbour (KNN) classifier.
The classifier will be trained on a number of neighbours, starting from min.num.neighbours, and increased gradually by 1 until max.num.neighbours is reached.
For each number of neighbour, the accuracy of the classifier will be computed.
Data will be normalised to range between 0 and 1 before used in training the classifier.
The normalisation method used is MinMaxScaling method.}
}
\value{
The performance of the classifier on different number of neighbours as well as some description on how the training was performed
i.e. what sampling process is used (how the data is split for training and testing),
the sample sizes (number of data points used for training/testing).
Included as well is the recommended number of neighbours for the data and the reasoning behind why that number of neighbours is the best.
}
\description{
train.knn.classifier - ...
}
