% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.hdf5.R
\name{write.hdf5}
\alias{write.hdf5}
\title{write.hdf5}
\usage{
write.hdf5(
  dat,
  channels = NULL,
  q.max = 0.99,
  flip.y = TRUE,
  random.crop.x = NULL,
  random.crop.y = NULL,
  random.crop.x.seed = 42,
  random.crop.y.seed = 21,
  print.spatial.plot = TRUE,
  chunk.size = NULL,
  compression = 0,
  plots = FALSE
)
}
\arguments{
\item{dat}{NO DEFAULT. Spatial data list}

\item{channels}{DEFAULT = NULL. Channels to include. If NULL, all channels included.}

\item{q.max}{DEFAULT = 0.99}

\item{flip.y}{DEFAULT = TRUE}

\item{random.crop.x}{DEFAULT = NULL}

\item{random.crop.y}{DEFAULT = NULL}

\item{random.crop.x.seed}{DEFAULT = 42}

\item{random.crop.y.seed}{DEFAULT = 21}

\item{print.spatial.plot}{DEFAULT = TRUE}

\item{chunk.size}{DEFAULT = NULL}

\item{compression}{DEFAULT = 0}

\item{plots}{DEFAULT = FALSE}

\item{merge.channels}{DEFAULT = NULL. Channels to use to create an additional 'merged' channel.}
}
\description{
write.hdf5
}
