% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.sumtable-percent.pos.R
\name{write.sumtable.percent.pos}
\alias{write.sumtable.percent.pos}
\title{write.sumtable.percent.pos}
\usage{
write.sumtable.percent.pos(x, sample.name, cluster.name, Markers, Cutoff)
}
\description{
write.sumtable.percent.pos
}
