% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.sumtables.R
\name{write.sumtables}
\alias{write.sumtables}
\title{write.sumtables}
\usage{
make.sumtables(x, ...)
}
\arguments{
\item{x}{NO DEFAULT. A data.table containing cells (rows) vs features/markers (columns). One column must represent sample names, and another must represent populations/clusters.}

\item{sample.col}{NO DEFAULT. Character. Name of the sample column (e.g. "Sample").}

\item{pop.col}{NO DEFAULT. Character. Name of the population/cluster column (e.g. "Population", "Cluster").}

\item{measure.col}{NO DEFAULT. A character or numeric vector indicating the columns to be measured (e.g. cellular columns -- c("CD45", "CD3e") etc).}

\item{annot.col}{DEFAULTS TO NULL. A character or numeric vector indicating the columns to be included as annotation columns (e.g. cellular columns -- c("Batch", "Group") etc). If groups are present, this column must be present here also.}

\item{group.col}{DEFAULTS TO NULL. Character. Which column represent groups (e.g. "Group"). This is for dividing data within the function. If you wish for groups to be included in the annotations, you will need to also include it in the annot.col argument.}

\item{do.frequencies}{DEFAULTS TO TRUE. Do you wish to create cell proportion and/or cell count results?}

\item{cell.counts}{DEFAULTS TO NULL. If you wish to generate cell.count results, a vector of cell counts (e.g. c(1000, 1500, 2439,)) representing the cell counts in each of the samples. Must be entered in the order the that unique sample names appear in the dataset.}

\item{do.mfi.per.sample}{DEFAULTS TO TRUE. Do you wish to generate MFI data (markers vs clusters) for each sample?}

\item{do.mfi.per.marker}{Currently inactive, set to FALSE.}

\item{mfi.type}{DEFAULTS TO "median". Can be "median" or "mean". Defines the type of function for calculating MFI data.}

\item{path}{DEFAULTS TO getwd(). Defines the directory for write CSV files.}
}
\value{
...
}
\description{
write.sumtables
}
\examples{
make.sumtables()

}
\references{
\url{https://sydneycytometry.org.au/spectre}.
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
