% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.spectre.R
\name{add.new.metadata}
\alias{add.new.metadata}
\title{Add new metadata to Spectre object}
\usage{
add.new.metadata(spectre_obj, metadata, metadata_name)
}
\arguments{
\item{spectre_obj}{A Spectre object.}

\item{metadata}{A data.table containing the metadata to add to the object.}

\item{metadata_name}{The name of the metadata.}
}
\value{
An updated Spectre object.
}
\description{
Add new metadata to the "other" slot in Spectre object
}
\examples{
dat <- create.spectre.object("cell_id")
metadata_dt <- data.table(marker=paste0("marker_", seq(1,10)), cofactors = seq(1,10))
dat <- add.new.metadata(dat, metadata_dt, "asinh_cofactors")
dat
}
