% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.spectre.R
\name{add.new.metadata}
\alias{add.new.metadata}
\title{Add new metadata to Spectre object}
\usage{
add.new.metadata(spectre_obj, metadata, metadata_name, dataset_name)
}
\arguments{
\item{spectre_obj}{A Spectre object.}

\item{metadata}{A list of metadata to add to the Spectre object.
Each element must be associated with a name.
For example, if one of the elements stores the cofactors used to do asinh transformation,
this list must be something like \code{list("cofactors" = 5)}.}

\item{dataset_name}{The name of the dataset to associate the metadata with.}
}
\value{
An updated Spectre object.
}
\description{
Add new information for a given dataset to the "metadata" slot in Spectre object.
Please see details for more information about how edge cases are handled.
}
\details{
If a metadata_name for the dataset_name already exists, it will be overwritten.

If dataset_name does not exist, i.e. you have either not added the dataset
associated with dataset_name or you have manually added the dataset but forgot
to update the metadata slot (i.e. not using add.new.data function),
it will produce an error.
}
\examples{
dat <- create.spectre.object("cell_id")
metadata_dt <- data.table(marker=paste0("marker_", seq(1,10)), cofactors = seq(1,10))
dat <- add.new.metadata(dat, metadata_dt, "asinh_cofactors")
dat
}
