% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.dt.R
\name{create.dt}
\alias{create.dt}
\title{create.dt - convert a Seurat or SingleCellExperiment to a data.table}
\usage{
create.dt(dat, from)
}
\arguments{
\item{dat}{NO DEFAULT. A Seurat or SingleCellExperiment or flowFrame object.}

\item{from}{DEFAULT = NULL. By default, the class of object will be detected
automatically, but this can be overwritten using from. Can be from =
'Seurat' or 'SingleCellExperiment' or 'flowFrame'.}
}
\value{
A list containing several elements.
The first element is a data.table which concatenates all components of dat into a single data.table.
The remaining elements contain the metadata of the dat such as the gene names,
cell barcodes, top x highly expressed genes, assays, etc.
}
\description{
This function converts a Seurat or SingleCellExperiment object into a list
containing a data.table, with vectors of gene and dimensionality reduction
}
\references{
\url{https://github.com/ImmuneDynamics/Spectre}.
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
