% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.dt.R
\name{create.dt}
\alias{create.dt}
\title{Create a data.table}
\usage{
create.dt(dat, from)
}
\arguments{
\item{dat}{NO DEFAULT. A Seurat or SingleCellExperiment or flowFrame object.}

\item{from}{DEFAULT = NULL. By default, the class of object will be detected
automatically, but this can be overwritten using from. Can be from =
'Seurat' or 'SingleCellExperiment' or 'flowFrame'.}
}
\value{
A list containing several elements.
The first element is a data.table which concatenates all components of dat into a single data.table.
The remaining elements contain the metadata of the dat such as the gene names,
cell barcodes, top x highly expressed genes, assays, etc.
}
\description{
Function to convert a Seurat or SingleCellExperiment or flowFrame object into a list
containing a data.table object storing the data, as well as a bunch of vectors grouping
the various features (columns) in the data.table, e.g. dim.reds vector will list
the kind of dimensionality reduction (by PCA or UMAP for instance) that have been applied
to the data and henceforth the columns in the data.table associated with them.
}
\references{
\url{https://immunedynamics.io/spectre/}.
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
Givanna Putri
}
