% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.clustered.R
\docType{data}
\name{demo.clustered}
\alias{demo.clustered}
\title{demo.clustered - Demo data.table with cells (rows) x markers (columns), with clusters, populations, and UMAP values added}
\format{
Demo data.table with cells (rows) x markers (columns), with clusters, populations, and UMAP values added.
}
\source{
Thomas M Ashhurst.
}
\usage{
demo.clustered
}
\description{
demo.clustered - Demo data.table with cells (rows) x markers (columns), with clusters, populations, and UMAP values added
}
\examples{
demo.clustered

}
\references{
\url{https://sydneycytometry.org.au/spectre}.
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
\keyword{datasets}
