% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.add.masks.R
\name{do.add.masks}
\alias{do.add.masks}
\title{do.add.masks}
\usage{
do.add.masks(
  spatial.dat,
  mask.loc,
  masks,
  mask.label = "cell_mask",
  correct.extent = TRUE,
  flip.y = TRUE,
  value.modifier = 65535,
  mask.ext = "_mask.tiff"
)
}
\description{
do.add.masks
}
