% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.extract.cell.dat.R
\name{do.extract.cell.dat}
\alias{do.extract.cell.dat}
\title{do.extract.cell.dat}
\usage{
do.extract.cell.dat(spatial.dat, target.dat)
}
\description{
This is a developmental Spectre-spatial function that is still in testing
phase with limited documentation. We recommend only using this function if
you know what you are doing.
}
