% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.extract.R
\name{do.extract}
\alias{do.extract}
\title{do.extract}
\usage{
do.extract(dat, mask, name = "CellData", fun = "mean")
}
\arguments{
\item{mask}{NO DEFAULT. Name of mask to use for cell data extraction}

\item{name}{DEFAULT = 'CellData'. Name of the extracted cell dataset.}

\item{fun}{DEFAULT = 'mean'. Summarisation function.}

\item{spatial.dat}{NO DEFAULT. Spatial data list}
}
\description{
do.extract
}
