% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.filter.percell.R
\name{do.filter.percell}
\alias{do.filter.percell}
\title{do.filter.percell}
\usage{
do.filter.percell(
  spatial.dat,
  per.cell,
  to,
  filter.by,
  id.col = "ObjectNumber",
  x.col = "Location_Center_X",
  y.col = "Location_Center_Y",
  simplify.cp.colname = TRUE,
  value.modifier = 65535
)
}
\description{
do.filter.percell
}
