% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.list.summary.R
\name{do.list.summary}
\alias{do.list.summary}
\title{Summarise the column names and row numbers for elements of a list}
\usage{
do.list.summary(dat)
}
\arguments{
\item{dat}{NO DEFAULT. A list of data.tables that you wish to create some summary information for}
}
\value{
Returns a new list summarising the data.tables in your list.
}
\description{
Summarise the column names and row numbers for elements of a list
}
\references{
\url{https://sydneycytometry.org.au/spectre}.
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
