% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.mn.R
\name{do.mn}
\alias{do.mn}
\title{do.mn - Multipurpose function for arcsinh transformation, quantile normalisation, noise reduction, and normalisation}
\usage{
do.mn(dat, use.cols)
}
\arguments{
\item{dat}{NO DEFAULT. A data.table}

\item{use.cols}{NO DEFAULT. Vector of column names that you wish to transform}

\item{asinh.cf}{DEFAULT = 15. The co-factor to use for arcsinh transformation}

\item{norm}{DEFAULT = c(0,1). New MIN and MAX values for each channel, as c(MIN, MAX).}

\item{lower.threshold}{DEFAULT = 0. Values below this number will be converted to this number.}

\item{upper.Q}{DEFAULT = 0.995. To not utilise the upper quantile function, simply use upper.Q = NULL or upper.Q = 1. When normalising, the new 'maximum' value (defined in the 'norm' argument, typically '1') will be set to this quantile target (for upper.Q = 0.995, this is the 99.5th percentile). As such, any values above the 99.5th percentile will increase above this new maximum value proportionally. Any target between 0 and 1 can be set.}

\item{append.name}{DEFAULT = '_transf'. Text to be appended to resultant column name}

\item{info.dir}{DEFAULT = getwd(). Directory address to save a .txt. file with the settings used for the MRN function.}
}
\value{
Returns a data.table with transformed data added as new columns.
}
\description{
This function facilitates multiple data transformation functions in a single process, including arcsinh transformation, noise reduction, and normalisation
}
\examples{
cell.dat <- do.mn()

}
\references{
Ashhurst, T. M., et al. (2019). \url{https://www.ncbi.nlm.nih.gov/pubmed/31077106}
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
