% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.mn.R
\name{do.mn}
\alias{do.mn}
\title{do.mn - Multipurpose function for arcsinh transformation, outlier management, and quantile-based re-scaling}
\usage{
do.mn(dat, use.cols)
}
\arguments{
\item{dat}{NO DEFAULT. A data.table}

\item{use.cols}{NO DEFAULT. Vector of column names that you wish to transform}

\item{asinh.cf}{DEFAULT = 15. The co-factor to use for arcsinh transformation}

\item{lower.Q}{DEFAULT = 0.001. Quantile target for values at the lower threshold}

\item{upper.Q}{DEFAULT = 0.995. Quantile target for rescaling, such that the data will be scaled between 0 (the minimum value after clipping) and 1 (upper.Q), where values above the upper.Q will increase above 1 proportionatly.}

\item{lower.func}{DEFAULT = 'clip'. Determine what is done to values below the lower.Q threshold. 'clip' means values below the lower.Q quantile threshold will be clipped to that value. NULL will skip.}

\item{append.name}{DEFAULT = '_transf'. Text to be appended to resultant column name}

\item{save.info}{DEFAULT = TRUE. Save's the transformation details to a .txt file.}

\item{info.dir}{DEFAULT = getwd(). Directory address to save a .txt. file with the settings used for the MRN function.}
}
\value{
Returns a data.table with transformed data added as new columns.
}
\description{
This function facilitates multiple data transformation functions in a single process, including arcsinh transformation, outlier management, and quantile-based re-scaling.
}
\examples{
cell.dat <- do.mn()

}
\references{
Ashhurst, T. M., et al. (2019). \url{https://www.ncbi.nlm.nih.gov/pubmed/31077106}
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
