% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.zscore.R
\name{do.zscore}
\alias{do.zscore}
\title{do.zscore - Calculate z.score for a dataset}
\usage{
do.zscore(dat, use.cols)
}
\arguments{
\item{dat}{NO DEFAULT. A data.table (or data.frame) containing the data to be converted to z-scores. Z-score transformed values will be added as new columns.}

\item{use.cols}{NO DEFAULT. The columns to be used for z-score calculations.}

\item{append.name}{DEFAULT = '_zscore'. Text to be appended to the end of the new z-score transformed columns.}

\item{replace}{DEFAULT = FALSE. If FALSE, appends new columns to the data.table. If TRUE, replaces the values in the existing columns with the z-score tranformed values, and does not change the column names.}
}
\value{
Returns a new data.table with z-score calculations for each selected column
}
\description{
do.zscore - Calculate z.score for a dataset
}
