% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.plot.R
\name{factor.plot}
\alias{factor.plot}
\title{factor.plot - Create a factor plot}
\usage{
factor.plot(x, ...)
}
\description{
This function allows you to plot cells on an XY plot, coloured by some factor (group, cluster, etc)
factor.plot is a wrapper around the heatmap.2 function from gplots.
}
