% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.files.R
\name{file.merge}
\alias{file.merge}
\title{file.merge - ...}
\usage{
file.merge(x, ...)
}
\arguments{
\item{x}{List of data frames or data tables}

\item{remove.duplicates}{Do you want to remove duplicates?}
}
\value{
A combined data.table
}
\description{
file.merge - ...
}
\examples{
file.merge()

}
\references{
\url{https://sydneycytometry.org.au/spectre}.
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
