% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_commonness_matrix_from_gdm.R
\name{generate_commonness_matrix_from_gdm}
\alias{generate_commonness_matrix_from_gdm}
\title{generate_commonness_matrix_from_gdm}
\usage{
generate_commonness_matrix_from_gdm(gdm_predictions, alpha_list)
}
\value{

}
\description{
Generates the best estimate objective (siteXsite commonness matrix) from predicted richness and predicted Bray-Curtis dissimilarity.
}
\details{
The function expects predicted per site richness as a vector, and Bray-Curtis dissimilarity as provided by the gdm-package (Fitzpatrick et al. 2020) output (as a list).
}
\examples{
example
}
