% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sorensen_matrix_from_gdm.R
\name{generate_sorensen_matrix_from_gdm}
\alias{generate_sorensen_matrix_from_gdm}
\title{generate_sorensen_matrix_from_gdm}
\usage{
generate_sorensen_matrix_from_gdm(gdm_predictions, n_sites)
}
\value{

}
\description{
Generate a sorensen siteXsite matrix from gdm-package output
}
\details{
details
}
\examples{
example
}
