% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled.factor.plot.R
\name{labelled.factor.plot}
\alias{labelled.factor.plot}
\title{labelled.factor.plot - ...}
\usage{
labelled.factor.plot(d, x.axis, y.axis, col.axis, title, dot.size,
  align.xy.by = NULL, align.col.by = NULL)
}
\description{
labelled.factor.plot - ...
}
