% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.autograph.R
\name{make.autograph}
\alias{make.autograph}
\title{make.autograph}
\usage{
make.autograph(
  x,
  x.axis,
  y.axis,
  colour.by,
  colours,
  y.axis.label,
  my_comparisons,
  title = paste0(y.axis.label, " - ", y.axis),
  filename = paste0(y.axis.label, " - ", y.axis, ".pdf"),
  path = getwd(),
  scale = "lin",
  dot.size = 5,
  width = 5,
  height = 5,
  y.first.level = 1.2,
  y.second.level = 1.5,
  Variance_test = "kruskal.test",
  Pairwise_test = "wilcox.test"
)
}
\description{
make.autograph
}
