% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.density.plot.R
\name{make.density.plot}
\alias{make.density.plot}
\title{Make a density plot (using ggpointdensity::geom_pointdensity())}
\usage{
make.density.plot(dat, x.axis, y.axis)
}
\arguments{
\item{dat}{NO DEFAULT. data.frame.}

\item{x.axis}{NO DEFAULT. Character. Column for X axis.}

\item{y.axis}{NO DEFAULT. Character. Column for Y axis.}

\item{colours}{DEFAULT = "viridis". Character. Colour scheme to use. Can be "spectral", "jet", "viridis", "magma", "inferno", or "BuPu".}

\item{align.xy.by}{DEFAULT = dat. data.table Sample to use to determine minimum and maximum X and Y axis values.}

\item{title}{DEFAULT = "Density". Character. Title for the plot.}

\item{dot.size}{DEFAULT = 1. Numeric. Size of the dots.}

\item{plot.width}{DEFAULT = 9. Width of the ggplot when saved to disk.}

\item{plot.height}{DEFAULT = 7. Height of the ggplot when saved to disk.}

\item{nudge_x}{DEFAULT = 0.5. When add.label = TRUE, distance the label is shifted from the centroid point on the X axis.}

\item{nudge_y}{DEFAULT = 0.5. When add.label = TRUE, distance the label is shifted from the centroid point on the Y axis.}

\item{square}{DEFAULT = TRUE. Ensures the plot is saved as a square. Set to FALSE if you want a plot with different X and Y lengths.}

\item{blank.axis}{DEFAULT = FALSE. Logical, do you want a minimalist graph?}

\item{save.to.disk}{DEFAULT = TRUE. Will save the ggplot to disk. If FALSE, will only show the ggplot.}

\item{path}{DEFAULT = getwd(). The location to save your ggplot. By default, will save to current working directory. Can be overidden.}
}
\description{
Method to create a density plot.
A density plot can summarise the cellular distribution across a given plot.
Can be used to plot any two markers.
Uses "ggplot2" for plots, "ggpointdensity" for adding density to plots, "RColorBrewer" for colour schemes.
}
\examples{
library(Spectre)
package.check()
package.load()

# Read data
cell.dat <- Spectre::demo.umap
cell.dat <- as.data.table(cell.dat)

# Draw plot
Spectre::make.density.plot(dat = cell.dat,
                          x.axis = "UMAP_42_X",
                          y.axis = "UMAP__42Y")

}
\author{
Felix Marsh-Wakefield, \email{felix.marsh-wakefield@sydney.edu.au}
}
