% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.density.plot.R
\name{make.density.plot}
\alias{make.density.plot}
\title{make.density.plot - ...}
\usage{
make.multi.marker.plot(dat, x.axis, y.axis, title, colours, dot.size, align.xy.by, align.col.by, save.to.disk, path, plot.width,  plot.height, blank.axis, ...)
}
\arguments{
\item{dat}{NO DEFAULT. data.frame.}

\item{x.axis}{NO DEFAULT. Character. Column for X axis.}

\item{y.axis}{NO DEFAULT. Character. Column for Y axis.}

\item{title}{DEFAULT = "Density". Character. Title for the plot.}

\item{colours}{DEFAULT = "viridis". Character. Colour scheme to use. Can be "spectral", "jet", "viridis", "magma", or "inferno".}

\item{dot.size}{DEFAULT = 1. Numeric. Size of the dots.}

\item{align.xy.by}{DEFAULT = dat. data.table Sample to use to determine minimum and maximum X and Y axis values.}

\item{align.col.by}{DEFAULT = dat. data.table. Sample to use to determine minimum and maximum colour values.}

\item{save.to.disk}{DEFAULT = TRUE. Will save the ggplot to disk. If FALSE, will only show the ggplot.}

\item{path}{DEFAULT = getwd(). The location to save your ggplot. By default, will save to current working directory. Can be overidden.}

\item{plot.width}{DEFAULT = 9. Width of the ggplot when saved to disk.}

\item{plot.height}{DEFAULT = 7. Height of the ggplot when saved to disk.}

\item{blank.axis}{DEFAULT = FALSE. Logical, do you want a minimalist graph?

This function will create a density plot for the chosen axes.}
}
\description{
make.density.plot - ...
}
