% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.multi.plot.R
\name{make.multi.plot}
\alias{make.multi.plot}
\title{Make multiple plots for each sample/group}
\usage{
make.multi.plot(dat, x.axis, y.axis, col.axis, type, plot.by, align.xy.by, align.col.by, colours, figure.title, dot.size, col.min.threshold, col.max.threshold, path, plot.width,  plot.height, blank.axis, save.each.plot)
}
\arguments{
\item{dat}{NO DEFAULT. A data frame containing all the data you wish to plot}

\item{x.axis}{NO DEFAULT. X axis}

\item{y.axis}{NO DEFAULT. Y axis}

\item{col.axis}{NO DEFAULT. Colour axis}

\item{type}{NO DEFAULT. Specify whether col.axis is continious ('colour', for markers) or a factor ('factor', for clusters, sample, groups etc), 'labelled.factor' or 'density' (for density plots - note figure.title must be entered).}

\item{plot.by}{NO DEFAULT. Name of the FACTORS that the plots will be divided by (e.g. Sample, Group, Batch etc).}

\item{align.xy.by}{NO DEFAULT. Align X and Y to a dataset}

\item{align.col.by}{NO DEFAULT. Align colour to a dataset}

\item{colours}{DEFAULTS to 'spectral'. What colour scheme do you want to use. Only used if type = 'colour', ignored if type = 'factor'. Can be 'jet', 'spectral', 'viridis', 'inferno', or 'magma'.}

\item{figure.title}{DEFAULTS to paste0(plot.by, " ", col.axis). Figure title. NOTE, CURRENTLY DISABLED.}

\item{dot.size}{DEFAULTS to 1. Size of dots}

\item{col.min.threshold}{DEFAULTS to 0.01. Minimum threshold for colour scale.}

\item{col.max.threshold}{DEFAULTS to 0.995. Maximum threshold for colour scale.}

\item{path}{DEFAULTS TO getwd() -- i.e. the current working directory. Path to the desired output directory}

\item{plot.width}{DEFAULTS to 9.}

\item{plot.height}{DEFAULTS to 7.}

\item{blank.axis}{DEFAULT = FALSE. Logical, do you want a minimalist graph?}

\item{save.each.plot}{DEFAULT = FALSE. Do you want to save each plot?}
}
\description{
Method to create multiple plots for each marker.
This function allows you to create a grid of plots, where the cells are subsetted by a certain factor (e.g. one sample per plot).
These can then be coloured by a marker or by another factor (e.g Group).
Makes use of Spectre functions make.colour.plot and make.density.plot.
}
\examples{
# Create grid of plots on demonstration data
Spectre::make.multi.plots(dat = Spectre::demo.umap,
                          x.axis <- "UMAP_42_X",
                          y.axis <- "UMAP_42_Y",
                          col.axis = "Group",
                          type = "factor",
                          plot.by = "Sample",
                          align.xy.by = dat,
                          align.col.by = dat)

}
\author{
Thomas Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
Felix Marsh-Wakefield, \email{felix.marsh-wakefield@sydney.edu.au}
}
