% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.multi.plots.R
\name{make.multi.plot}
\alias{make.multi.plot}
\title{make.multi.plot}
\usage{
make.multi.plot(x, ...)
}
\arguments{
\item{dat}{NO DEFAULT. A data frame containing all the data you wish to plot}

\item{x.axis}{NO DEFAULT. X axis}

\item{y.axis}{NO DEFAULT. Y axis}

\item{col.axis}{NO DEFAULT. Colour axis}

\item{type}{NO DEFAULT. Specify whether col.axis is continious ('colour', for markers) or a factor ('factor', for clusters, sample, groups etc), 'labelled.factor' or 'density' (for density plots - note figure.title must be entered).}

\item{plot.by}{NO DEFAULT. Name of the FACTORS that the plots will be divided by (e.g. Sample, Group, Batch etc).}

\item{align.xy.by}{NO DEFAULT. Align X and Y to a dataset}

\item{align.col.by}{NO DEFAULT. Align colour to a dataset}

\item{colours}{DEFAULTS to 'spectral'. What colour scheme do you want to use. Only used if type = 'colour', ignored if type = 'factor'. Can be 'jet', 'spectral', 'viridis', 'inferno', or 'magma'.}

\item{figure.title}{DEFAULTS to paste0(plot.by, " ", col.axis). Figure title. NOTE, CURRENTLY DISABLED.}

\item{dot.size}{DEFAULTS to 1. Size of dots}

\item{col.min.threshold}{DEFAULTS to 0.01. Minimum threshold for colour scale.}

\item{col.max.threshold}{DEFAULTS to 0.995. Maximum threshold for colour scale.}

\item{path}{DEFAULTS TO getwd() -- i.e. the current working directory. Path to the desired output directory}

\item{plot.width}{DEFAULTS to 9.}

\item{plot.height}{DEFAULTS to 7.}

\item{blank.axis}{DEFAULT = FALSE. Logical, do you want a minimalist graph?}

\item{save.each.plot}{DEFAULT = FALSE. Do you want to save each plot?

This function allows you to create a grid of plots, where the cells are subsetted by a certain factor (e.g. one sample per plot). These can then be coloured by a marker or by another factor (e.g Group).}
}
\description{
make.multi.plot
}
