% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.pheatmap.R
\name{make.pheatmap}
\alias{make.pheatmap}
\title{make.pheatmap - Create a 'pretty' heatmap (pheatmap)}
\usage{
make.pheatmap(dat, ...)
}
\arguments{
\item{dat}{NO DEFAULT. data.frame. Clusters/populations vs markers (MFI) or Clusters/populations vs samples or (MFI or cell numbers). No default.}

\item{file.name}{NO DEFAULT. Character. What do you want to call the file, including the extension.}

\item{plot.title}{NO DEFAULT. Character.}

\item{sample.col}{NO DEFAULT. Character. Specify the name of the column that indicates samples. No default.}

\item{plot.cols}{NO DEFAULT. Character vector. Name of columns you wish to plot on the heatmap.}

\item{annot.cols}{Character. Columns which contain values that you do NOT want to plot in the heatmap, e.g. sample name, group name, Live/Dead etc. No default.}

\item{transpose}{Logical. Do you want to transpose the heatmap. Defaults to FALSE.}

\item{normalise}{Logical. Only applies to standard heatmaps (i.e. when is.fold = FALSE). TRUE to normalise each column between 0 and 1, FALSE to plot the raw values. Defaults to TRUE.}

\item{is.fold}{Logical. TRUE for fold-change heatmap, FALSE for normal values heatmap. Defaults to FALSE.}

\item{fold.range}{Numeric vector. For fold-change heatmaps, what is the maxium and minimum values that should be plotted. Example: for a max of 3, and a minimum of -3 would b c(3,-3). Defaults to NULL (which will use the max and min within the dataset)}

\item{dendrograms}{Character. Do you want to include dendrograms on columns/rows. Can be "both", "row", "column", or "none. Defaults to "both.}

\item{n.row.groups}{Numeric.}

\item{n.col.groups}{Numeric.}

\item{row.sep}{Numeric. Only used if not clustering rows.}

\item{col.sep}{Numeric. Only used if not clustering columns}

\item{cell.size}{Numeric. Defaults to 15.}

\item{standard.colours}{Character. DEFAULTS to "YlGnBu", can also be "viridis", "magma", "inferno", "spectral".}

\item{fold.colours}{Character.}

\item{colour.scheme}{Character. Only one option currently.}

\item{y.margin}{Numeric.}

\item{x.margin}{Numeric.}
}
\value{
prints a 'pretty' heatmap with dendrograms.
}
\description{
This function allows you to create a coloured heatmap for clusters vs markers (coloured by MFI) or clusters vs samples (coloured by MFI or cell counts). Can be set to plot 'fold-change' values that are provided in log2.
make.pheatmap is a wrapper around the pheatmap function from pheatmap.
}
\examples{
make.pheatmap()

}
\references{
\url{https://sydneycytometry.org.au/spectre}. Helpful examples at \url{https://davetang.org/muse/2018/05/15/making-a-heatmap-in-r-with-the-pheatmap-package/}
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
