% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.spatial.plot.R
\name{make.spatial.plot}
\alias{make.spatial.plot}
\title{make.spatial.plot - Create a dot plot (X vs Y) coloured by a selected continuous column (e.g. marker expression)}
\usage{
make.spatial.plot(
  spatial.dat,
  roi,
  roi.marker,
  cell.dat,
  cell.colour,
  cell.x = "X",
  cell.y = "Y",
  add.outlines = TRUE,
  flip.y.axis = TRUE,
  cell.colour.type = "numeric",
  raster.mask.size = 0.1,
  raster.mask.colour = "gold",
  raster.min.threshold = 0,
  raster.max.threshold = 0.99,
  title = paste0(roi),
  col.min.threshold = 0.01,
  col.max.threshold = 0.995,
  colours = "spectral",
  dot.size = 1.5,
  dot.alpha = 0.6,
  align.xy.by = cell.dat,
  align.col.by = cell.dat,
  save.to.disk = TRUE,
  path = getwd(),
  plot.width = 9,
  plot.height = 7,
  blank.axis = FALSE
)
}
\description{
make.spatial.plot - Create a dot plot (X vs Y) coloured by a selected continuous column (e.g. marker expression)
}
