% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.volcano.plot.R
\name{make.volcano.plot}
\alias{make.volcano.plot}
\title{Make an enhanced volcano plot}
\usage{
make.volcano.plot(
  dat.p,
  dat.fc,
  vars,
  title,
  subtitle = NULL,
  xlim = c(-6, 8),
  ylim = c(0, 4),
  path = getwd(),
  width = 7,
  height = 8,
  ...
)
}
\description{
Make an enhanced volcano plot
}
