% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.check.R
\name{package.check}
\alias{package.check}
\title{package.check - a function to check the installation of all required packages.}
\usage{
package.check(type = "general")
}
\arguments{
\item{type}{DEFAULT = "general". If "general", then checks for the packages required for general Spectre usage. If "spatial", then checks for additional packages required for spatial analysis. If "ML", then checks for additional packages required for machine-learing functionality.}
}
\value{
returns an error message if one of the common use packages are not installed. Proceeds in order of package importance, and only the first error message encountered will be returned.
}
\description{
This function allows you to check to see if all the common use packages dependencies for Spectre are installed.
See \url{https://sydneycytometry.org.au/spectre} for usage instructions and vignettes.
}
\examples{
package.check()

}
\references{
\url{https://sydneycytometry.org.au/spectre}
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
