% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_commonness.R
\name{plot_commonness}
\alias{plot_commonness}
\title{plot_commonness}
\usage{
plot_commonness(species_grid, target)
}
\arguments{
\item{species_grid}{Optimized grid using run_optimization.}

\item{target}{Pairwise matrix of species in common.}
}
\value{
ggplot
}
\description{
Plot commonness
}
\details{
Plot a heatmap of commonnes between observed data and optimized data.
}
\references{
xxx
}
