% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_commonness.R
\name{plot_commonness}
\alias{plot_commonness}
\title{plot_commonness}
\usage{
plot_commonness(x, target)
}
\arguments{
\item{x}{Results object of run_optimization_min_conf()}

\item{target}{Pairwise matrix of species in common.}
}
\value{
ggplot
}
\description{
Plot commonness
}
\details{
Plot a heatmap of commonness between observed data and optimized data.
}
\references{
xxx
}
