% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.spatial.files.R
\name{read.spatial.files}
\alias{read.spatial.files}
\title{read.spatial.files - read in TIFF files into a spatial data object (list of rasters)}
\usage{
read.spatial.files(
  roi.loc,
  rois,
  mask.loc = NULL,
  masks = NULL,
  mask.ext = NULL,
  cell.dat = NULL,
  image.dat = NULL,
  panel.dat = NULL,
  ext = ".tiff",
  read.tiff.stacks = TRUE,
  convert.to.raster = TRUE,
  correct.extent = TRUE,
  value.modifier = 65535,
  calculate.polygons = TRUE,
  calculate.outlines = TRUE,
  calculate.centroids = TRUE
)
}
\description{
read.spatial.files - read in TIFF files into a spatial data object (list of rasters)
}
