% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.spatial.files.R
\name{read.spatial.files}
\alias{read.spatial.files}
\title{Read TIFF files into R and and create spatial data object}
\usage{
read.spatial.files()
}
\arguments{
\item{dir}{Directory location of where the ROI files are located}

\item{correct.extent}{Correct extent so that the minimum is 0,0. Default TRUE}

\item{flip.y}{Flip the data arrangement for the y-axis. Default TRUE}

\item{value.modifier}{Data modifier based on image processing. Default 65535}

\item{lois}{Default = NULL. You can provide a list of only some of the ROI folders to read in.}
}
\value{
Returns a list of spatial data objects.
}
\description{
Read TIFF files into R and and create spatial data object
}
\examples{

}
