% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.chronoclust.R
\name{run.chronoclust}
\alias{run.chronoclust}
\title{run.chronoclust - Function to run Chronoclust. 
ChronoClust was originally written in Python 3. It'll be run using Reticulate.
Please run run.prepare.chronoclust before running this function to
properly setup the python environment for running chronoclust.}
\usage{
run.chronoclust(data.files, output.dir, config)
}
\arguments{
\item{data.files}{NO DEFAULT. A vector containing the location of csv file containing the time series datasets. 
Make sure that each 1 csv file is data for only 1 time point. 
Thus if you have 3 time points, this vector have 3 elements in it.}

\item{output.dir}{NO DEFAULT. Character. The directory for Chronoclust to write its result to.}

\item{config}{DEFAULT NULL. List of Character. Specify the parameter for ChronoClust as a list. If none is given, ChronoClust will
run on a default config. Only specify the parameter that you do not wish to use the default value for.}
}
\description{
run.chronoclust - Function to run Chronoclust. 
ChronoClust was originally written in Python 3. It'll be run using Reticulate.
Please run run.prepare.chronoclust before running this function to
properly setup the python environment for running chronoclust.
}
\examples{
run.chronoclust(data.files=c("day1.csv","day2.csv"), output.dir="outdir", config=list(beta= 0.2, delta= 0.05, epsilon= 0.03, mu= 0.01))

}
\references{
\url{https://sydneycytometry.org.au/spectre}.
}
\author{
Givanna Putri, \email{ghar1821@uni.sydney.edu.au}
}
