% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.chronoclust.R
\name{run.chronoclust}
\alias{run.chronoclust}
\title{Run the ChronoClust algorithm}
\usage{
run.chronoclust(data.files, output.dir, config = NULL)
}
\arguments{
\item{data.files}{A character vector. Path and name of the csv files storing the time series dataset. \strong{1 csv file per time point}.}

\item{output.dir}{Character. Path to the directory where the clustering results will be stored.}

\item{config}{A named numeric list. List containing the value of ChronoClust's hyper-parameter. The name of the element must correspond to one of ChronoClust's parameter name such as epsilon, upsilon, etc. The numeric value must correspond to the value assigned for the corresponding parameter. 
\emph{Only include parameters that you want to override.} Those you prefer to set to default value need not be included in the list.}
}
\description{
Method to run ChronoClust clustering algorithm.
For this to work properly, a Python session must already be embedded within the currently running R session.
This is because ChronoClust is originally written in Python 3. 
Spectre includes a function \code{\link{run.prepare.chronoclust}} that uses Reticulate to prepare and embed Python session within R session.
}
\examples{
# Prepare Python session
run.prepare.chronoclust(environment_name = "chronoclust-R", 
                        create_environment = TRUE, 
                        install_dependencies = TRUE)

# Specify the data files location and output directory
data.files <- c("cc_example/day1.csv", "cc_example/day2.csv")
output.dir <- "cc_example/output"

# Specify the parameters that need to be overriden
config <- list(beta = 0.2, 
               delta = 0.05, 
               epsilon = 0.03, 
               mu = 0.01)

run.chronoclust(data.files = data.files, 
                output.dir = output.dir, 
                config = config)

}
\seealso{
\url{https://github.com/ghar1821/Chronoclust} for ChronoClust's Python implementation.

\code{\link{run.prepare.chronoclust}} for how to embed Python session within R session.
}
\author{
Givanna Putri, \email{ghar1821@uni.sydney.edu.au}
}
