% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.chronoclust.R
\name{run.chronoclust}
\alias{run.chronoclust}
\title{run.chronoclust - ...}
\usage{
run.chronoclust(x, ...)
}
\arguments{
\item{config.xml.file.location}{NO DEFAULT. Location of xml file describing the parameter values for Chronoclust to operate on.}

\item{input.xml.file.location}{NO DEFAULT. Location of xml file describing the location of the dataset (for each timepoint) for Chronoclust to operate on.}

\item{output.dir.location}{NO DEFAULT. The directory for Chronoclust to write its result to.}

\item{conda.env.name}{NO DEFAULT. The conda virtual environment name where Chronoclust wil run.


This function run a time-series clustering and cluster tracking algorithm Chronoclust.
ChronoClust was originally written in Python. It'll be run using Reticulate.
Only anaconda/miniconda virtual environment is supported by this function. This environment need to be setup before running the function.
Please visit the various guide in the following link to install anaconda/miniconda and setup an environment:
https://docs.conda.io/projects/conda/en/latest/user-guide/index.html
ChronoClust requires the following python package to run: numpy, pandas, scikit-learn, tqdm, numba.
You can either install them separately after setting up your anaconda environment 
or just simply import them from the prepared environment script provided in our github repository (worked_example/Chronoclust_workflow)}
}
\description{
run.chronoclust - ...
}
