% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.cycombine.R
\name{run.cycombine}
\alias{run.cycombine}
\title{Run cyCombine}
\usage{
run.cycombine(
  dat,
  use_cols,
  batch_col,
  cell_id_col,
  xdim = 8,
  ydim = 8,
  rlen = 10,
  parametric = TRUE,
  seed = 42,
  covar = NULL,
  anchor = NULL,
  norm_method = "scale",
  ties_method = "average",
  verbose = TRUE
)
}
\arguments{
\item{dat}{NO DEFAULT.
A Spectre object containing the data to apply batch correction to.}

\item{use_cols}{NO DEFAULT.
A vector of character column names to apply batch correction to.}

\item{batch_col}{NO default.
The column that denotes the batch or dataset that each cell belongs to}

\item{cell_id_col}{Character. The column in \code{dat} denoting
the unique identifier of the cells.}

\item{xdim}{DEFAULT 8. X-dimension size of the SOM grid.}

\item{ydim}{DEFAULT 8. Y-dimension size of the SOM grid.}

\item{rlen}{DEFAULT 10. Number of times the data is presented to the SOM network.
Higher values are recommended if 10 does not appear to perform well.}

\item{parametric}{DEFAULT TRUE.
If TRUE, the parametric version of ComBat is used.
If FALSE, the non-parametric version is used.}

\item{seed}{DEFAULT 42. Seed to use when creating the SOM.}

\item{covar}{DEFAULT NULL. The covariate ComBat should use.
Can be a vector or a column name in the Spectre object (element data_source).
If NULL, no covar will be used}

\item{anchor}{DEFAULT NULL.
Experimental: A column or vector specifying which samples are replicates and
which are not.
If specified, this column will be used as a covariate in ComBat.
Be aware that it may be confounded with the condition.}

\item{norm_method}{DEFAULT scale.
Normalization method. Should be either 'rank', 'scale' or 'qnorm'.}

\item{ties_method}{DEFAULT average.
The method to handle ties, when using rank.
See ?rank for other options.}

\item{verbose}{DEFAULT TRUE
Whether to print progress messages. TRUE to print, FALSE to suppress.}
}
\value{
batch corrected data.table
}
\description{
Run cyCombine algorithm to do batch correction.
}
\references{
Pedersen, C.B., Dam, S.H., Barnkob, M.B. et al.
cyCombine allows for robust integration of single-cell cytometry datasets
within and across technologies.
Nat Commun 13, 1698 (2022).
https://doi.org/10.1038/s41467-022-29383-5

Ashhurst, T.M., Marsh‐Wakefield, F., Putri, G.H. et al.
Integration, exploration, and analysis of high‐dimensional single‐cell
cytometry data using Spectre.
Cytometry Part A, 101(3), pp.237-253 (2022).
https://doi.org/10.1002/cyto.a.24350
}
