% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.flowsom.R
\name{run.flowsom}
\alias{run.flowsom}
\title{Run the FlowSOM algorithm}
\usage{
run.flowsom(dat, use.cols, meta.k, xdim, ydim, clust.seed, meta.seed, clust.name, meta.clust.name)
}
\arguments{
\item{dat}{NO DEFAULT. data.frame. Input sample.}

\item{use.cols}{NO DEFAULT. Vector of column names to use for clustering.}

\item{meta.k}{DEFAULT = 20. Numeric. Number of clusters to create. If set to zero (0), no metaclusters will be created.}

\item{xdim}{DEFAULT = 10. Numeric. Number of first level clusters across the x-axis. xdim x ydim = total number of first level FlowSOM clusters.}

\item{ydim}{DEFAULT = 10. Numeric. Number of first level clusters across the y-axis. xdim x ydim = total number of first level FlowSOM clusters.}

\item{clust.seed}{DEFAULT = 42 Numeric. Clustering seed for reproducibility.}

\item{meta.seed}{DEFAULT = 42 Numeric. Metaclustering seed for reproducibility.}

\item{clust.name}{DEFAULT = "FlowSOM_cluster". Character. Name of the resulting 'cluster' parameter.}

\item{meta.clust.name}{DEFAULT = "FlowSOM_metacluster". Character. Name of the resulting 'metacluster' parameter.}

\item{mem.ctrl}{DEFAULT = TRUE. Runs gc() (garbage collection) after a number of steps to free up memory that hasn't been released quickly enough.}
}
\description{
Method to run the FlowSOM clustering algorithm.
This function runs FlowSOM on a data.table with cells (rows) vs markers (columns) with new columns for FlowSOM clusters and metaclusters.
Output data will be "flowsom.res.original" (for clusters) and "flowsom.res.meta" (for metaclusters).
Uses the R packages "FlowSOM" for clustering, "flowCore" for handling .fcs files, "Biobase" for creating a flow frame, "data.table" for handling data.table format.
}
\examples{
# Run FlowSOM on demonstration dataset
Spectre::run.flowsom(Spectre::demo.start,
                     use.cols = c(5:6,8:9,11:13,16:19,21:30,32))

}
\author{
Thomas Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
Felix Marsh-Wakefield, \email{felix.marsh-wakefield@sydney.edu.au}
}
