% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.flowsom.R
\name{run.flowsom}
\alias{run.flowsom}
\title{run.flowsom - ...}
\usage{
run.flowsom(x, ...)
}
\arguments{
\item{x}{data.frame. Input sample. No default.}

\item{clustering.cols}{Vector of column names to use for clustering. It is possible to use a vector of column numbers here but this is not recommended, as No default.}

\item{meta.k}{Numeric. Number of clusters to create. DEFAULT = 20.}

\item{xdim}{Numeric. Number of first level clusters across the x-axis. xdim x ydim = total number of first level FlowSOM clusters. DEFAULT = 10.}

\item{ydim}{Numeric. Number of first level clusters across the y-axis. xdim x ydim = total number of first level FlowSOM clusters. DEFAULT = 10.}

\item{clust.seed}{Numeric. Clustering seed for reproducibility. DEFAULT = 42}

\item{meta.seed}{Numeric. Metaclustering seed for reproducibility. DEFAULT = 42.}

\item{clust.name}{Character. Name of the resulting 'cluster' parameter. DEFAULT = "FlowSOM_cluster".}

\item{meta.clust.name}{Character. Name of the resulting 'metacluster' parameter. DEFAULT = "FlowSOM_metacluster".

This function runs FlowSOM on a dataframe with cells (rows) vs markers (columns), and returns 'res' with result columns}
}
\description{
run.flowsom - ...
}
