% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.flowsom.R
\name{run.flowsom}
\alias{run.flowsom}
\title{run.flowsom - ...}
\usage{
run.flowsom(x, ...)
}
\arguments{
\item{x}{data.frame. Input sample. No default.}

\item{meta.k}{Numeric. Number of clusters to create. No default.}

\item{clustering.cols}{Vector of column names to use for clustering. No default.}

\item{clust.seed}{Numeric. Clustering seed for reproducibility. No default.}

\item{meta.seed}{Numeric. Metaclustering seed for reproducibility. No default.}

\item{clust.name}{Character. Name of the resulting 'cluster' parameter. Defaults to "FlowSOM_cluster".}

\item{meta.clust.name}{Character. Name of the resulting 'metacluster' parameter. Defaults to "FlowSOM_metacluster".

This function runs FlowSOM on a dataframe with cells (rows) vs markers (columns), and returns 'res' with result columns}
}
\description{
run.flowsom - ...
}
