% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.harmony.R
\name{run.harmony}
\alias{run.harmony}
\title{Run Harmony batch alignment algorithm}
\usage{
run.harmony(
  dat,
  cell_id_col,
  use_cols,
  batch_col,
  do_pca = TRUE,
  npcs = 20,
  theta = NULL,
  lambda = 1,
  sigma = 0.1,
  nclust = NULL,
  max_iter = 10,
  early_stop = TRUE,
  ncores = 1,
  plot_convergence = FALSE,
  return_object = FALSE,
  verbose = FALSE,
  seed = 42,
  alpha = 0.2,
  tau = 0,
  block.size = 0.05,
  max.iter.cluster = 200,
  epsilon.cluster = 0.001,
  epsilon.harmony = 0.01
)
}
\arguments{
\item{dat}{NO DEFAULT. A data.table.}

\item{cell_id_col}{Character. The column in \code{dat} denoting
the unique identifier of the cells.}

\item{use_cols}{NO DEFAULT.
A vector of character column names to apply batch correction to.}

\item{batch_col}{NO default. The column that denotes the batch or dataset that each cell belongs to}

\item{do_pca}{DEFAULT TRUE. Whether to perform PCA on the data.}

\item{npcs}{DEFAULT 20. If doing PCA on the data, the number of PCs to compute.}

\item{theta}{DEFAULT NULL. Diversity clustering penalty parameter. Specify for each
variable in vars_use. theta=0 does not encourage any
diversity. Larger values of theta result in more diverse clusters.}

\item{lambda}{DEFAULT NULL. Ridge regression penalty parameter.
Specify for each batch.
Default lambda=1. Lambda must be strictly positive. Smaller values result
in more aggressive correction.}

\item{sigma}{DEFAULT 0.1.
Width of soft kmeans clusters. Default sigma=0.1. Sigma scales
the distance from a cell to cluster centroids. Larger values of sigma
result in cells assigned to more clusters. Smaller values of sigma make
soft kmeans cluster approach hard clustering.}

\item{nclust}{DEFAULT NULL. Number of clusters in model.
nclust=1 equivalent to simple linear regression.}

\item{max_iter}{DEFAULT 10.
Maximum number of rounds to run Harmony.
One round of Harmony involves one clustering and one correction step.}

\item{early_stop}{DEFAULT TRUE.
Enable early stopping for harmony.
The harmonization process will stop when the change of
objective function between corrections drops below 1e-4.
Only used if 'epsilon.harmony' is set to NULL,}

\item{ncores}{DEFAULT 1.
Number of processors to be used for math operations when optimized BLAS is available.
If BLAS is not supporting multithreaded then this option has no effect.
By default, ncores=1 which runs as a single-threaded process.
Although Harmony supports multiple cores,
it is not optimized for multithreading.
Increase this number for large datasets iff single-core performance is not
adequate and optimized BLAS is available.}

\item{plot_convergence}{DEFAULT FALSE.
Whether to print the convergence plot of the clustering objective function.
TRUE to plot, FALSE to suppress.
This can be useful for debugging.}

\item{return_object}{DEFAULT FALSE.
(Advanced Usage) Whether to return the Harmony object or only the corrected PCA embeddings.
If TRUE, this will be stored in the metadata slot.}

\item{verbose}{DEFAULT FALSE.
Whether to print progress messages. TRUE to print, FALSE to suppress.}

\item{seed}{DEFAULT 42. Seed used when harmony runs PCA.}

\item{alpha}{DEFAULT 0.2. Advanced Harmony parameter.
When setting lambda = NULL and use lambda estimation mode,
lambda would be determined by the expected number of cells assuming
independence between batches and clusters.
i.e., lambda = alpha * expected number of cells.
Alpha should be 0 < alpha < 1}

\item{tau}{DEFAULT 0. Protection against overclustering small datasets with
large ones. 'tau' is the expected number of cells per cluster.}

\item{block.size}{DEFAULT 0.05. Advanced Harmony parameter.
What proportion of cells to update during clustering. Between 0 to 1.
Larger values may be faster but less accurate.}

\item{max.iter.cluster}{DEFAULT 20. Advanced Harmony parameter.
Maximum number of rounds to run clustering at each round of Harmony.}

\item{epsilon.cluster}{DEFAULT 0.001. Advanced Harmony parameter.
Convergence tolerance for clustering round of Harmony.
Set to -Inf to never stop early.}

\item{epsilon.harmony}{DEFAULT 0.01. Advanced Harmony parameter.
Convergence tolerance for Harmony.
Set to -Inf to never stop early.
When 'epsilon.harmony' is set to not NULL,
then user-supplied values of 'early_stop' is ignored.}
}
\value{
Returns a data.table with aligned data added in new columns.
}
\description{
Use Harmony to do batch correction.
For more details on Harmony see:
https://portals.broadinstitute.org/harmony/articles/quickstart.html.
}
\references{
Korsunsky, I., Millard, N., Fan, J. et al.
Fast, sensitive and accurate integration of single-cell data with Harmony.
Nat Methods 16, 1289–1296 (2019).
https://doi.org/10.1038/s41592-019-0619-0

Ashhurst, T.M., Marsh‐Wakefield, F., Putri, G.H. et al.
Integration, exploration, and analysis of high‐dimensional single‐cell
cytometry data using Spectre.
Cytometry Part A, 101(3), pp.237-253 (2022).
https://doi.org/10.1002/cyto.a.24350
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}

Givanna Putri
}
