% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.pca.R
\name{run.pca}
\alias{run.pca}
\title{run.pca - ...}
\usage{
run.pca(x, use.cols, ...)
}
\arguments{
\item{x}{data.frame. No default.}

\item{use.cols}{Vector of numbers, reflecting the columns to use for clustering. No default.}

\item{cor}{A logical value indicating whether the calculation should use the correlation matrix or the covariance matrix. (The correlation matrix can only be used if there are no constant variables.). Default = TRUE.}

\item{scores}{A logical value indicating whether the score on each principal component should be calculated. Default = TRUE.}

\item{scree.plot}{Option to create scree plots. Note this will require the input of an elbow point during run. Will save generated scree plot. Default = TRUE.}

\item{component.loading}{Option to create plots for each component. Requires scree.plot = TRUE. Default = TRUE.}

\item{marker.contribution}{Option to create plot showing the contribution of each marker. Horizontal red line represents the average marker contribution if all markers contributed equally. Requires scree.plot = TRUE. Default = TRUE.}

\item{loading.plot}{Option to create scree plots. Will save generated loading plot. Default = TRUE.}

\item{individual.samples}{Option to run above plots on a per sample basis. Only samples that have a cell number greater than the number of parameters/markers will be included. Default = FALSE.}

\item{sample.code}{Parameter to define column that will differentiate between samples. Must be quoted. Requires scree.plot = TRUE.  Used for "individual.samples".Default = "FileName"}

\item{top.tally}{Number of top markers that contributed to PCA across each sample. Used for "individual.samples". Defalut = 10.}

\item{path}{Location to save plots. Default = getwd() (working directory)

This function runs a principal component analysis (PCA) on a dataframe with cells (rows) vs markers (columns), returning chosen figures. Uses the base R package "stats" for PCA, "factoextra" for scree and loading plots, "data.table" for saving .csv files, "ggplot2" for saving plots, "gtools" for rearranging data order, "dplyr" for selecting top n values.}
}
\description{
run.pca - ...
}
