% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.phenograph.R
\name{run.phenograph}
\alias{run.phenograph}
\title{run.phenograph - Run phenograph clustering}
\usage{
run.phenograph()
}
\arguments{
\item{ref.dat}{NO DEFAULT. A data.table consisting of the 'refernece' data you will use to train the alignment algorithm}
}
\value{
Returns a data.table with clustering results added
}
\description{
This function allows you to perform phenograph clustering on a data.table
}
\examples{

cell.dat <- Spectre::demo.asinh
cell.dat <- Spectre::do.subsample(cell.dat, 5000)

use.cols <- names(cell.dat)[c(11:19)]

cell.dat <- run.phenograph(dat = cell.dat, use.cols = use.cols)


}
\references{
Ashhurst, T. M., et al. (2019). \url{https://www.ncbi.nlm.nih.gov/pubmed/31077106}
}
\author{
Thomas M Ashhurst, \email{thomas.ashhurst@sydney.edu.au}
}
