% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.tsne.R
\name{run.tsne}
\alias{run.tsne}
\title{run.tsne - ...}
\usage{
run.tsne(dat, use.cols, tsne.x.name, tsne.y.name, tsne.seed, dims, initial_dims, perplexity, theta, check_duplicates, pca, max_iter, verbose, is_distance, Y_init, stop_lying_iter, mom_switch_iter, momentum, final_momentum, eta, exaggeration_factor, ...)
}
\arguments{
\item{dat}{NO DEFAULT. data.frame.}

\item{use.cols}{NO DEFAULT. Vector of numbers, reflecting the columns to use for dimensionality reduction.}

\item{tsne.x.name}{DEFAULT = "tSNE_X". Character. Name of tSNE x-axis.}

\item{tsne.y.name}{DEFAULT = "tSNE_Y". Character. Name of tSNE y-axis.}

\item{tsne.seed}{DEFAULT = 42. Numeric. Seed value for reproducibility.}

\item{dims}{DEFAULT = 2. Number of dimensions for output results, either 2 or 3.}

\item{initial_dims}{DEFAULT = 50. Number of dimensions retained in initial PCA step.}

\item{perplexity}{DEFAULT = 30.}

\item{theta}{DEFAULT = 0.5. Use 0.5 for Barnes-Hut tSNE, 0.0 for exact tSNE (takes longer).}

\item{check_duplicates}{DEFAULT = FALSE.}

\item{pca}{DEFAULT = TRUE. Runs PCA prior to tSNE run.}

\item{max_iter}{DEFAULT = 1000. Maximum number of iterations.}

\item{verbose}{DEFAULT = TRUE.}

\item{is_distance}{DEFAULT = FALSE. Experimental, using X as a distance matrix.}

\item{Y_init}{DEFAULT = NULL. Recommend NULL for random initialisation.}

\item{stop_lying_iter}{DEFAULT = 250. Number of iterations of early exaggeration.}

\item{mom_switch_iter}{DEFAULT = 250. Number of iterations before increased momentum of spread.}

\item{momentum}{DEFAULT = 0.5. Initial momentum of spread.}

\item{final_momentum}{DEFAULT = 0.8. Momentum of spread at 'final_momentum'.}

\item{eta}{DEFAULT = 200. Learning rate.}

\item{exaggeration_factor}{DEFAULT = 12.0. Factor used during early exaggeration.

This function runs tSNE on a dataframe with cells (rows) vs markers (columns), and returns 'res' with result columns. Uses the Rtsne package. For more information on parameter choices and effects, check out https://distill.pub/2016/misread-tsne/.}
}
\description{
run.tsne - ...
}
