% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.umap.R
\name{run.umap}
\alias{run.umap}
\title{run.umap - ...}
\usage{
run.umap(x, ...)
}
\arguments{
\item{x}{NO DEFAULT. Input data.table or data.frame.}

\item{use.cols}{NO DEFAULT. Vector of column names or numbers for clustering.}

\item{umap.seed}{DEFAULT = 42. Numeric. Seed value for reproducibility.}

\item{suffix}{DEFAULT = blank.

This function runs UMAP on a dataframe with cells (rows) vs markers (columns), and returns 'res' with result columns.}
}
\description{
run.umap - ...
}
