% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_optimization.R
\name{run_optimization}
\alias{run_optimization}
\title{run_optimization}
\usage{
run_optimization(alpha_list, total_gamma, target, max_runs,
  annealing = 0, energy_threshold, patience, verbose = TRUE)
}
\arguments{
\item{alpha_list}{Matrix of predicted alpha diversity in each cell.}

\item{total_gamma}{Total (estimated) species in the system.}

\item{target}{Pairwise matrix of species in common.}

\item{max_runs}{Max number of loops before stopping.}

\item{annealing}{Probability to keep modified matrix even if energy increased.}

\item{energy_threshold}{Optimization stops if energy threshold is reached.}

\item{patience}{Number of runs with no improvement before stopping.}

\item{verbose}{It TRUE, progress report is printed}
}
\value{
Best matrix of common species between each site.
}
\description{
xxx
}
\details{
This is the function which runs the optimization algorithm.
There are a number of steps involved
\enumerate{
\item generate initial solution site x species matrix (only needs to be run once per landscape)
\item alter solution site x species matrix and select best site x site matrix based on Dsol
\item Run optimization algorithm until some stop requirement is met
}
}
\references{
xxx
}
